@@ -58,45 +58,71 @@
   },
   "dependencies": {
     "classnames": "^2.2.4",
-    "create-react-class": "^15.5.2",
     "prop-types": "^15.5.8",
-    "react-input-autosize": "^1.1.3"
+    "react-input-autosize": "^1.1.5"
   },
   "description": "A Select control built with and for ReactJS",
   "devDependencies": {
-    "babel": "^5.8.23",
-    "babel-eslint": "^4.1.3",
-    "chai": "^3.5.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-external-helpers": "^6.22.0",
+    "babel-plugin-istanbul": "^4.1.4",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
     "coveralls": "^2.11.12",
-    "eslint": "^1.10.3",
-    "eslint-plugin-react": "^3.15.0",
-    "gulp": "^3.9.1",
+    "create-react-class": "^15.5.2",
+    "cross-env": "^5.0.5",
+    "css-loader": "^0.28.7",
+    "eslint": "^4.6.1",
+    "eslint-plugin-react": "^7.3.0",
+    "extract-text-webpack-plugin": "^3.0.0",
+    "html-loader": "^0.5.1",
+    "html-webpack-plugin": "^2.30.1",
     "isomorphic-fetch": "^2.2.1",
     "istanbul": "^0.4.5",
-    "jsdom": "^9.5.0",
+    "jsdom": "^9.12.0",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.5",
+    "less-plugin-clean-css": "^1.5.1",
     "mocha": "^3.0.2",
-    "react": "^15.0",
-    "react-addons-shallow-compare": "^15.0",
-    "react-addons-test-utils": "^15.0",
-    "react-component-gulp-tasks": "^0.7.7",
-    "react-dom": "^15.0",
+    "nps": "^5.7.1",
+    "nps-utils": "^1.3.0",
+    "nyc": "^11.1.0",
+    "react": "^15.5.0",
+    "react-addons-shallow-compare": "^15.5.0",
+    "react-dom": "^15.5.0",
     "react-gravatar": "^2.4.5",
-    "react-highlight-words": "^0.3.0",
-    "react-virtualized": "^7.23.0",
-    "react-virtualized-select": "^1.4.0",
-    "sinon": "^1.17.5",
-    "unexpected": "^10.17.0",
-    "unexpected-dom": "^3.1.0",
-    "unexpected-react": "^3.2.4",
-    "unexpected-sinon": "^10.4.0"
+    "react-highlight-words": "^0.8.1",
+    "react-test-renderer": "^15.6.1",
+    "react-virtualized": "^9.9.0",
+    "react-virtualized-select": "^3.1.0",
+    "rollup": "^0.49.2",
+    "rollup-plugin-babel": "^3.0.2",
+    "rollup-plugin-commonjs": "^8.2.0",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-plugin-uglify": "^2.0.1",
+    "sinon": "^3.2.1",
+    "style-loader": "^0.18.2",
+    "uglify-es": "^3.0.28",
+    "unexpected": "^10.35.0",
+    "unexpected-dom": "^4.0.0",
+    "unexpected-react": "^4.1.0",
+    "unexpected-sinon": "^10.4.0",
+    "webpack": "^3.5.5",
+    "webpack-dev-server": "^2.7.1"
   },
   "directories": {},
   "dist": {
-    "shasum": "9d316f252b1adc372ddb5cdf1f119c6b7cfdb5d6",
-    "tarball": "https://registry.npmjs.org/react-select/-/react-select-1.0.0-rc.5.tgz"
+    "shasum": "5a33618559fcb3824929dde71cfe1f406aa9a0df",
+    "tarball": "https://registry.npmjs.org/react-select/-/react-select-1.0.0-rc.6.tgz"
   },
-  "gitHead": "1d15068d054993ffa73b3a06766d9b6079b3fe9a",
+  "gitHead": "c4830c6c85a0564fb82d3069feee9ba839bfb5bb",
   "homepage": "https://github.com/JedWatson/react-select#readme",
+  "jsnext:main": "dist/react-select.es.js",
   "keywords": [
     "combobox",
     "form",
@@ -122,8 +148,8 @@
   "name": "react-select",
   "optionalDependencies": {},
   "peerDependencies": {
-    "react": "^0.14 || ^15.0.0-rc || ^15.0",
-    "react-dom": "^0.14 || ^15.0.0-rc || ^15.0"
+    "react": "^0.14 || ^15.0 || ^16.0.0-rc || ^16.0",
+    "react-dom": "^0.14 || ^15.0 || ^16.0.0-rc || ^16.0"
   },
   "readme": "ERROR: No README data found!",
   "repository": {
