@@ -1,65 +1,45 @@
-'use strict';
-
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+import React from 'react';
+import Select from './Select';
+import Async from './Async';
+import Creatable from './Creatable';
 
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
-
-var _Select = require('./Select');
-
-var _Select2 = _interopRequireDefault(_Select);
-
-function reduce(obj) {
-	var props = arguments.length <= 1 || arguments[1] === undefined ? {} : arguments[1];
-
-	return Object.keys(obj).reduce(function (props, key) {
-		var value = obj[key];
+function reduce(obj, props = {}) {
+	return Object.keys(obj).reduce((props, key) => {
+		const value = obj[key];
 		if (value !== undefined) props[key] = value;
 		return props;
 	}, props);
 }
 
-var AsyncCreatable = (0, _createReactClass2['default'])({
-	displayName: 'AsyncCreatableSelect',
+class AsyncCreatableSelect extends React.Component {
 
-	focus: function focus() {
+	focus() {
 		this.select.focus();
-	},
+	}
 
-	render: function render() {
-		var _this = this;
-
-		return _react2['default'].createElement(
-			_Select2['default'].Async,
+	render() {
+		return React.createElement(
+			Async,
 			this.props,
-			function (asyncProps) {
-				return _react2['default'].createElement(
-					_Select2['default'].Creatable,
-					_this.props,
-					function (creatableProps) {
-						return _react2['default'].createElement(_Select2['default'], _extends({}, reduce(asyncProps, reduce(creatableProps, {})), {
-							onInputChange: function (input) {
-								creatableProps.onInputChange(input);
-								return asyncProps.onInputChange(input);
-							},
-							ref: function (ref) {
-								_this.select = ref;
-								creatableProps.ref(ref);
-								asyncProps.ref(ref);
-							}
-						}));
+			asyncProps => React.createElement(
+				Creatable,
+				this.props,
+				creatableProps => React.createElement(Select, _extends({}, reduce(asyncProps, reduce(creatableProps, {})), {
+					onInputChange: input => {
+						creatableProps.onInputChange(input);
+						return asyncProps.onInputChange(input);
+					},
+					ref: ref => {
+						this.select = ref;
+						creatableProps.ref(ref);
+						asyncProps.ref(ref);
 					}
-				);
-			}
+				}))
+			)
 		);
 	}
-});
+};
 
-module.exports = AsyncCreatable;
+export default AsyncCreatableSelect;
