@@ -1,85 +1,28 @@
+var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
+
+function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
+
 /*!
   Copyright (c) 2016 Jed Watson.
   Licensed under the MIT License (MIT), see
   http://jedwatson.github.io/react-select
 */
+import React from 'react';
+import PropTypes from 'prop-types';
+import ReactDOM from 'react-dom';
+import AutosizeInput from 'react-input-autosize';
+import classNames from 'classnames';
 
-'use strict';
+import defaultArrowRenderer from './utils/defaultArrowRenderer';
+import defaultFilterOptions from './utils/defaultFilterOptions';
+import defaultMenuRenderer from './utils/defaultMenuRenderer';
+import defaultClearRenderer from './utils/defaultClearRenderer';
 
-Object.defineProperty(exports, '__esModule', {
-	value: true
-});
-
-var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
-
-function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
-
-var _propTypes = require('prop-types');
-
-var _propTypes2 = _interopRequireDefault(_propTypes);
-
-var _reactDom = require('react-dom');
-
-var _reactDom2 = _interopRequireDefault(_reactDom);
-
-var _reactInputAutosize = require('react-input-autosize');
-
-var _reactInputAutosize2 = _interopRequireDefault(_reactInputAutosize);
-
-var _classnames = require('classnames');
-
-var _classnames2 = _interopRequireDefault(_classnames);
-
-var _utilsDefaultArrowRenderer = require('./utils/defaultArrowRenderer');
-
-var _utilsDefaultArrowRenderer2 = _interopRequireDefault(_utilsDefaultArrowRenderer);
-
-var _utilsDefaultFilterOptions = require('./utils/defaultFilterOptions');
-
-var _utilsDefaultFilterOptions2 = _interopRequireDefault(_utilsDefaultFilterOptions);
-
-var _utilsDefaultMenuRenderer = require('./utils/defaultMenuRenderer');
-
-var _utilsDefaultMenuRenderer2 = _interopRequireDefault(_utilsDefaultMenuRenderer);
-
-var _utilsDefaultClearRenderer = require('./utils/defaultClearRenderer');
-
-var _utilsDefaultClearRenderer2 = _interopRequireDefault(_utilsDefaultClearRenderer);
-
-var _Async = require('./Async');
-
-var _Async2 = _interopRequireDefault(_Async);
-
-var _AsyncCreatable = require('./AsyncCreatable');
-
-var _AsyncCreatable2 = _interopRequireDefault(_AsyncCreatable);
-
-var _Creatable = require('./Creatable');
-
-var _Creatable2 = _interopRequireDefault(_Creatable);
-
-var _Option = require('./Option');
-
-var _Option2 = _interopRequireDefault(_Option);
-
-var _Value = require('./Value');
-
-var _Value2 = _interopRequireDefault(_Value);
+import Option from './Option';
+import Value from './Value';
 
 function stringifyValue(value) {
-	var valueType = typeof value;
+	const valueType = typeof value;
 	if (valueType === 'string') {
 		return value;
 	} else if (valueType === 'object') {
@@ -91,182 +34,90 @@ function stringifyValue(value) {
 	}
 }
 
-var stringOrNode = _propTypes2['default'].oneOfType([_propTypes2['default'].string, _propTypes2['default'].node]);
+const stringOrNode = PropTypes.oneOfType([PropTypes.string, PropTypes.node]);
 
-var instanceId = 1;
+let instanceId = 1;
 
-var Select = (0, _createReactClass2['default'])({
+class Select extends React.Component {
 
-	displayName: 'Select',
+	constructor(props) {
+		super(props);
 
-	propTypes: {
-		addLabelText: _propTypes2['default'].string, // placeholder displayed when you want to add a label on a multi-value input
-		'aria-describedby': _propTypes2['default'].string, // HTML ID(s) of element(s) that should be used to describe this input (for assistive tech)
-		'aria-label': _propTypes2['default'].string, // Aria label (for assistive tech)
-		'aria-labelledby': _propTypes2['default'].string, // HTML ID of an element that should be used as the label (for assistive tech)
-		arrowRenderer: _propTypes2['default'].func, // Create drop-down caret element
-		autoBlur: _propTypes2['default'].bool, // automatically blur the component when an option is selected
-		autofocus: _propTypes2['default'].bool, // autofocus the component on mount
-		autosize: _propTypes2['default'].bool, // whether to enable autosizing or not
-		backspaceRemoves: _propTypes2['default'].bool, // whether backspace removes an item if there is no text input
-		backspaceToRemoveMessage: _propTypes2['default'].string, // Message to use for screenreaders to press backspace to remove the current item - {label} is replaced with the item label
-		className: _propTypes2['default'].string, // className for the outer element
-		clearAllText: stringOrNode, // title for the "clear" control when multi: true
-		clearRenderer: _propTypes2['default'].func, // create clearable x element
-		clearValueText: stringOrNode, // title for the "clear" control
-		clearable: _propTypes2['default'].bool, // should it be possible to reset value
-		deleteRemoves: _propTypes2['default'].bool, // whether backspace removes an item if there is no text input
-		delimiter: _propTypes2['default'].string, // delimiter to use to join multiple values for the hidden field value
-		disabled: _propTypes2['default'].bool, // whether the Select is disabled or not
-		escapeClearsValue: _propTypes2['default'].bool, // whether escape clears the value when the menu is closed
-		filterOption: _propTypes2['default'].func, // method to filter a single option (option, filterString)
-		filterOptions: _propTypes2['default'].any, // boolean to enable default filtering or function to filter the options array ([options], filterString, [values])
-		ignoreAccents: _propTypes2['default'].bool, // whether to strip diacritics when filtering
-		ignoreCase: _propTypes2['default'].bool, // whether to perform case-insensitive filtering
-		inputProps: _propTypes2['default'].object, // custom attributes for the Input
-		inputRenderer: _propTypes2['default'].func, // returns a custom input component
-		instanceId: _propTypes2['default'].string, // set the components instanceId
-		isLoading: _propTypes2['default'].bool, // whether the Select is loading externally or not (such as options being loaded)
-		joinValues: _propTypes2['default'].bool, // joins multiple values into a single form field with the delimiter (legacy mode)
-		labelKey: _propTypes2['default'].string, // path of the label value in option objects
-		matchPos: _propTypes2['default'].string, // (any|start) match the start or entire string when filtering
-		matchProp: _propTypes2['default'].string, // (any|label|value) which option property to filter on
-		menuBuffer: _propTypes2['default'].number, // optional buffer (in px) between the bottom of the viewport and the bottom of the menu
-		menuContainerStyle: _propTypes2['default'].object, // optional style to apply to the menu container
-		menuRenderer: _propTypes2['default'].func, // renders a custom menu with options
-		menuStyle: _propTypes2['default'].object, // optional style to apply to the menu
-		multi: _propTypes2['default'].bool, // multi-value input
-		name: _propTypes2['default'].string, // generates a hidden <input /> tag with this field name for html forms
-		noResultsText: stringOrNode, // placeholder displayed when there are no matching search results
-		onBlur: _propTypes2['default'].func, // onBlur handler: function (event) {}
-		onBlurResetsInput: _propTypes2['default'].bool, // whether input is cleared on blur
-		onChange: _propTypes2['default'].func, // onChange handler: function (newValue) {}
-		onClose: _propTypes2['default'].func, // fires when the menu is closed
-		onCloseResetsInput: _propTypes2['default'].bool, // whether input is cleared when menu is closed through the arrow
-		onFocus: _propTypes2['default'].func, // onFocus handler: function (event) {}
-		onInputChange: _propTypes2['default'].func, // onInputChange handler: function (inputValue) {}
-		onInputKeyDown: _propTypes2['default'].func, // input keyDown handler: function (event) {}
-		onMenuScrollToBottom: _propTypes2['default'].func, // fires when the menu is scrolled to the bottom; can be used to paginate options
-		onOpen: _propTypes2['default'].func, // fires when the menu is opened
-		onValueClick: _propTypes2['default'].func, // onClick handler for value labels: function (value, event) {}
-		openAfterFocus: _propTypes2['default'].bool, // boolean to enable opening dropdown when focused
-		openOnFocus: _propTypes2['default'].bool, // always open options menu on focus
-		optionClassName: _propTypes2['default'].string, // additional class(es) to apply to the <Option /> elements
-		optionComponent: _propTypes2['default'].func, // option component to render in dropdown
-		optionRenderer: _propTypes2['default'].func, // optionRenderer: function (option) {}
-		options: _propTypes2['default'].array, // array of options
-		pageSize: _propTypes2['default'].number, // number of entries to page when using page up/down keys
-		placeholder: stringOrNode, // field placeholder, displayed when there's no value
-		required: _propTypes2['default'].bool, // applies HTML5 required attribute when needed
-		resetValue: _propTypes2['default'].any, // value to use when you clear the control
-		scrollMenuIntoView: _propTypes2['default'].bool, // boolean to enable the viewport to shift so that the full menu fully visible when engaged
-		searchable: _propTypes2['default'].bool, // whether to enable searching feature or not
-		simpleValue: _propTypes2['default'].bool, // pass the value to onChange as a simple value (legacy pre 1.0 mode), defaults to false
-		style: _propTypes2['default'].object, // optional style to apply to the control
-		tabIndex: _propTypes2['default'].string, // optional tab index of the control
-		tabSelectsValue: _propTypes2['default'].bool, // whether to treat tabbing out while focused to be value selection
-		value: _propTypes2['default'].any, // initial field value
-		valueComponent: _propTypes2['default'].func, // value component to render
-		valueKey: _propTypes2['default'].string, // path of the label value in option objects
-		valueRenderer: _propTypes2['default'].func, // valueRenderer: function (option) {}
-		wrapperStyle: _propTypes2['default'].object },
+		this.handleTouchOutside = this.handleTouchOutside.bind(this);
+		this.handleTouchMove = this.handleTouchMove.bind(this);
+		this.handleTouchStart = this.handleTouchStart.bind(this);
+		this.handleTouchEnd = this.handleTouchEnd.bind(this);
+		this.handleTouchEndClearValue = this.handleTouchEndClearValue.bind(this);
+		this.handleMouseDown = this.handleMouseDown.bind(this);
+		this.handleMouseDownOnArrow = this.handleMouseDownOnArrow.bind(this);
+		this.handleMouseDownOnMenu = this.handleMouseDownOnMenu.bind(this);
+		this.handleInputFocus = this.handleInputFocus.bind(this);
+		this.handleInputBlur = this.handleInputBlur.bind(this);
+		this.handleInputChange = this.handleInputChange.bind(this);
+		this.handleInputValueChange = this.handleInputValueChange.bind(this);
+		this.handleKeyDown = this.handleKeyDown.bind(this);
+		this.handleValueClick = this.handleValueClick.bind(this);
+		this.handleMenuScroll = this.handleMenuScroll.bind(this);
+		this.handleRequired = this.handleRequired.bind(this);
+		this.getOptionLabel = this.getOptionLabel.bind(this);
+		this.onOptionRef = this.onOptionRef.bind(this);
+		this.clearValue = this.clearValue.bind(this);
+		this.removeValue = this.removeValue.bind(this);
+		this.selectValue = this.selectValue.bind(this);
+		this.focusOption = this.focusOption.bind(this);
 
-	// optional style to apply to the component wrapper
-	statics: { Async: _Async2['default'], AsyncCreatable: _AsyncCreatable2['default'], Creatable: _Creatable2['default'] },
-
-	getDefaultProps: function getDefaultProps() {
-		return {
-			addLabelText: 'Add "{label}"?',
-			arrowRenderer: _utilsDefaultArrowRenderer2['default'],
-			autosize: true,
-			backspaceRemoves: true,
-			backspaceToRemoveMessage: 'Press backspace to remove {label}',
-			clearable: true,
-			clearAllText: 'Clear all',
-			clearRenderer: _utilsDefaultClearRenderer2['default'],
-			clearValueText: 'Clear value',
-			deleteRemoves: true,
-			delimiter: ',',
-			disabled: false,
-			escapeClearsValue: true,
-			filterOptions: _utilsDefaultFilterOptions2['default'],
-			ignoreAccents: true,
-			ignoreCase: true,
-			inputProps: {},
-			isLoading: false,
-			joinValues: false,
-			labelKey: 'label',
-			matchPos: 'any',
-			matchProp: 'any',
-			menuBuffer: 0,
-			menuRenderer: _utilsDefaultMenuRenderer2['default'],
-			multi: false,
-			noResultsText: 'No results found',
-			onBlurResetsInput: true,
-			onCloseResetsInput: true,
-			optionComponent: _Option2['default'],
-			pageSize: 5,
-			placeholder: 'Select...',
-			required: false,
-			scrollMenuIntoView: true,
-			searchable: true,
-			simpleValue: false,
-			tabSelectsValue: true,
-			valueComponent: _Value2['default'],
-			valueKey: 'value'
-		};
-	},
-
-	getInitialState: function getInitialState() {
-		return {
+		this.state = {
 			inputValue: '',
 			isFocused: false,
 			isOpen: false,
 			isPseudoFocused: false,
 			required: false
 		};
-	},
+	}
 
-	componentWillMount: function componentWillMount() {
+	componentWillMount() {
 		this._instancePrefix = 'react-select-' + (this.props.instanceId || ++instanceId) + '-';
-		var valueArray = this.getValueArray(this.props.value);
+		const valueArray = this.getValueArray(this.props.value);
 
 		if (this.props.required) {
 			this.setState({
 				required: this.handleRequired(valueArray[0], this.props.multi)
 			});
 		}
-	},
+	}
 
-	componentDidMount: function componentDidMount() {
+	componentDidMount() {
 		if (this.props.autofocus) {
 			this.focus();
 		}
-	},
+	}
 
-	componentWillReceiveProps: function componentWillReceiveProps(nextProps) {
-		var valueArray = this.getValueArray(nextProps.value, nextProps);
+	componentWillReceiveProps(nextProps) {
+		const valueArray = this.getValueArray(nextProps.value, nextProps);
 
 		if (nextProps.required) {
 			this.setState({
 				required: this.handleRequired(valueArray[0], nextProps.multi)
 			});
+		} else if (this.props.required) {
+			// Used to be required but it's not any more
+			this.setState({ required: false });
 		}
-	},
+	}
 
-	componentWillUpdate: function componentWillUpdate(nextProps, nextState) {
+	componentWillUpdate(nextProps, nextState) {
 		if (nextState.isOpen !== this.state.isOpen) {
 			this.toggleTouchOutsideEvent(nextState.isOpen);
-			var handler = nextState.isOpen ? nextProps.onOpen : nextProps.onClose;
+			const handler = nextState.isOpen ? nextProps.onOpen : nextProps.onClose;
 			handler && handler();
 		}
-	},
+	}
 
-	componentDidUpdate: function componentDidUpdate(prevProps, prevState) {
+	componentDidUpdate(prevProps, prevState) {
 		// focus to the selected option
 		if (this.menu && this.focused && this.state.isOpen && !this.hasScrolledToOption) {
-			var focusedOptionNode = _reactDom2['default'].findDOMNode(this.focused);
-			var menuNode = _reactDom2['default'].findDOMNode(this.menu);
+			let focusedOptionNode = ReactDOM.findDOMNode(this.focused);
+			let menuNode = ReactDOM.findDOMNode(this.menu);
 			menuNode.scrollTop = focusedOptionNode.offsetTop;
 			this.hasScrolledToOption = true;
 		} else if (!this.state.isOpen) {
@@ -275,8 +126,8 @@ var Select = (0, _createReactClass2['default'])({
 
 		if (this._scrollToFocusedOptionOnUpdate && this.focused && this.menu) {
 			this._scrollToFocusedOptionOnUpdate = false;
-			var focusedDOM = _reactDom2['default'].findDOMNode(this.focused);
-			var menuDOM = _reactDom2['default'].findDOMNode(this.menu);
+			var focusedDOM = ReactDOM.findDOMNode(this.focused);
+			var menuDOM = ReactDOM.findDOMNode(this.menu);
 			var focusedRect = focusedDOM.getBoundingClientRect();
 			var menuRect = menuDOM.getBoundingClientRect();
 			if (focusedRect.bottom > menuRect.bottom || focusedRect.top < menuRect.top) {
@@ -293,17 +144,17 @@ var Select = (0, _createReactClass2['default'])({
 			this.setState({ isFocused: false }); // eslint-disable-line react/no-did-update-set-state
 			this.closeMenu();
 		}
-	},
+	}
 
-	componentWillUnmount: function componentWillUnmount() {
+	componentWillUnmount() {
 		if (!document.removeEventListener && document.detachEvent) {
 			document.detachEvent('ontouchstart', this.handleTouchOutside);
 		} else {
 			document.removeEventListener('touchstart', this.handleTouchOutside);
 		}
-	},
+	}
 
-	toggleTouchOutsideEvent: function toggleTouchOutsideEvent(enabled) {
+	toggleTouchOutsideEvent(enabled) {
 		if (enabled) {
 			if (!document.addEventListener && document.attachEvent) {
 				document.attachEvent('ontouchstart', this.handleTouchOutside);
@@ -317,54 +168,54 @@ var Select = (0, _createReactClass2['default'])({
 				document.removeEventListener('touchstart', this.handleTouchOutside);
 			}
 		}
-	},
+	}
 
-	handleTouchOutside: function handleTouchOutside(event) {
+	handleTouchOutside(event) {
 		// handle touch outside on ios to dismiss menu
 		if (this.wrapper && !this.wrapper.contains(event.target)) {
 			this.closeMenu();
 		}
-	},
+	}
 
-	focus: function focus() {
+	focus() {
 		if (!this.input) return;
 		this.input.focus();
-	},
+	}
 
-	blurInput: function blurInput() {
+	blurInput() {
 		if (!this.input) return;
 		this.input.blur();
-	},
+	}
 
-	handleTouchMove: function handleTouchMove(event) {
+	handleTouchMove(event) {
 		// Set a flag that the view is being dragged
 		this.dragging = true;
-	},
+	}
 
-	handleTouchStart: function handleTouchStart(event) {
+	handleTouchStart(event) {
 		// Set a flag that the view is not being dragged
 		this.dragging = false;
-	},
+	}
 
-	handleTouchEnd: function handleTouchEnd(event) {
+	handleTouchEnd(event) {
 		// Check if the view is being dragged, In this case
 		// we don't want to fire the click event (because the user only wants to scroll)
 		if (this.dragging) return;
 
 		// Fire the mouse events
 		this.handleMouseDown(event);
-	},
+	}
 
-	handleTouchEndClearValue: function handleTouchEndClearValue(event) {
+	handleTouchEndClearValue(event) {
 		// Check if the view is being dragged, In this case
 		// we don't want to fire the click event (because the user only wants to scroll)
 		if (this.dragging) return;
 
 		// Clear the value
 		this.clearValue(event);
-	},
+	}
 
-	handleMouseDown: function handleMouseDown(event) {
+	handleMouseDown(event) {
 		// if the event was triggered by a mousedown and not the primary
 		// button, or if the component is disabled, ignore it.
 		if (this.props.disabled || event.type === 'mousedown' && event.button !== 0) {
@@ -381,6 +232,7 @@ var Select = (0, _createReactClass2['default'])({
 
 		// for the non-searchable select, toggle the menu
 		if (!this.props.searchable) {
+			// TODO: This code means that if a select is searchable, onClick the options menu will not appear, only on subsequent click will it open.
 			this.focus();
 			return this.setState({
 				isOpen: !this.state.isOpen
@@ -393,7 +245,7 @@ var Select = (0, _createReactClass2['default'])({
 			// Call focus() again here to be safe.
 			this.focus();
 
-			var input = this.input;
+			let input = this.input;
 			if (typeof input.getInput === 'function') {
 				// Get the actual DOM input if the ref is an <AutosizeInput /> component
 				input = input.getInput();
@@ -409,12 +261,12 @@ var Select = (0, _createReactClass2['default'])({
 			});
 		} else {
 			// otherwise, focus the input and open the menu
-			this._openAfterFocus = true;
+			this._openAfterFocus = this.props.openAfterFocus;
 			this.focus();
 		}
-	},
+	}
 
-	handleMouseDownOnArrow: function handleMouseDownOnArrow(event) {
+	handleMouseDownOnArrow(event) {
 		// if the event was triggered by a mousedown and not the primary
 		// button, or if the component is disabled, ignore it.
 		if (this.props.disabled || event.type === 'mousedown' && event.button !== 0) {
@@ -429,9 +281,9 @@ var Select = (0, _createReactClass2['default'])({
 		event.preventDefault();
 		// close the menu
 		this.closeMenu();
-	},
+	}
 
-	handleMouseDownOnMenu: function handleMouseDownOnMenu(event) {
+	handleMouseDownOnMenu(event) {
 		// if the event was triggered by a mousedown and not the primary
 		// button, or if the component is disabled, ignore it.
 		if (this.props.disabled || event.type === 'mousedown' && event.button !== 0) {
@@ -442,26 +294,25 @@ var Select = (0, _createReactClass2['default'])({
 
 		this._openAfterFocus = true;
 		this.focus();
-	},
+	}
 
-	closeMenu: function closeMenu() {
+	closeMenu() {
 		if (this.props.onCloseResetsInput) {
 			this.setState({
 				isOpen: false,
 				isPseudoFocused: this.state.isFocused && !this.props.multi,
-				inputValue: ''
+				inputValue: this.handleInputValueChange('')
 			});
 		} else {
 			this.setState({
 				isOpen: false,
-				isPseudoFocused: this.state.isFocused && !this.props.multi,
-				inputValue: this.state.inputValue
+				isPseudoFocused: this.state.isFocused && !this.props.multi
 			});
 		}
 		this.hasScrolledToOption = false;
-	},
+	}
 
-	handleInputFocus: function handleInputFocus(event) {
+	handleInputFocus(event) {
 		if (this.props.disabled) return;
 		var isOpen = this.state.isOpen || this._openAfterFocus || this.props.openOnFocus;
 		if (this.props.onFocus) {
@@ -472,9 +323,9 @@ var Select = (0, _createReactClass2['default'])({
 			isOpen: isOpen
 		});
 		this._openAfterFocus = false;
-	},
+	}
 
-	handleInputBlur: function handleInputBlur(event) {
+	handleInputBlur(event) {
 		// The check for menu.contains(activeElement) is necessary to prevent IE11's scrollbar from closing the menu in certain contexts.
 		if (this.menu && (this.menu === document.activeElement || this.menu.contains(document.activeElement))) {
 			this.focus();
@@ -490,20 +341,16 @@ var Select = (0, _createReactClass2['default'])({
 			isPseudoFocused: false
 		};
 		if (this.props.onBlurResetsInput) {
-			onBlurredState.inputValue = '';
+			onBlurredState.inputValue = this.handleInputValueChange('');
 		}
 		this.setState(onBlurredState);
-	},
+	}
 
-	handleInputChange: function handleInputChange(event) {
-		var newInputValue = event.target.value;
+	handleInputChange(event) {
+		let newInputValue = event.target.value;
 
-		if (this.state.inputValue !== event.target.value && this.props.onInputChange) {
-			var nextState = this.props.onInputChange(newInputValue);
-			// Note: != used deliberately here to catch undefined and null
-			if (nextState != null && typeof nextState !== 'object') {
-				newInputValue = '' + nextState;
-			}
+		if (this.state.inputValue !== event.target.value) {
+			newInputValue = this.handleInputValueChange(newInputValue);
 		}
 
 		this.setState({
@@ -511,9 +358,20 @@ var Select = (0, _createReactClass2['default'])({
 			isPseudoFocused: false,
 			inputValue: newInputValue
 		});
-	},
+	}
 
-	handleKeyDown: function handleKeyDown(event) {
+	handleInputValueChange(newValue) {
+		if (this.props.onInputChange) {
+			let nextState = this.props.onInputChange(newValue);
+			// Note: != used deliberately here to catch undefined and null
+			if (nextState != null && typeof nextState !== 'object') {
+				newValue = '' + nextState;
+			}
+		}
+		return newValue;
+	}
+
+	handleKeyDown(event) {
 		if (this.props.disabled) return;
 
 		if (typeof this.props.onInputKeyDown === 'function') {
@@ -595,30 +453,29 @@ var Select = (0, _createReactClass2['default'])({
 				return;
 		}
 		event.preventDefault();
-	},
+	}
 
-	handleValueClick: function handleValueClick(option, event) {
+	handleValueClick(option, event) {
 		if (!this.props.onValueClick) return;
 		this.props.onValueClick(option, event);
-	},
+	}
 
-	handleMenuScroll: function handleMenuScroll(event) {
+	handleMenuScroll(event) {
 		if (!this.props.onMenuScrollToBottom) return;
-		var target = event.target;
-
+		let { target } = event;
 		if (target.scrollHeight > target.offsetHeight && !(target.scrollHeight - target.offsetHeight - target.scrollTop)) {
 			this.props.onMenuScrollToBottom();
 		}
-	},
+	}
 
-	handleRequired: function handleRequired(value, multi) {
+	handleRequired(value, multi) {
 		if (!value) return true;
 		return multi ? value.length === 0 : Object.keys(value).length === 0;
-	},
+	}
 
-	getOptionLabel: function getOptionLabel(op) {
+	getOptionLabel(op) {
 		return op[this.props.labelKey];
-	},
+	}
 
 	/**
   * Turns a value into an array from the given options
@@ -626,92 +483,77 @@ var Select = (0, _createReactClass2['default'])({
   * @param	{Object}		nextProps	- optionally specify the nextProps so the returned array uses the latest configuration
   * @returns	{Array}	the value of the select represented in an array
   */
-	getValueArray: function getValueArray(value, nextProps) {
-		var _this = this;
-
+	getValueArray(value, nextProps) {
 		/** support optionally passing in the `nextProps` so `componentWillReceiveProps` updates will function as expected */
-		var props = typeof nextProps === 'object' ? nextProps : this.props;
+		const props = typeof nextProps === 'object' ? nextProps : this.props;
 		if (props.multi) {
 			if (typeof value === 'string') value = value.split(props.delimiter);
 			if (!Array.isArray(value)) {
 				if (value === null || value === undefined) return [];
 				value = [value];
 			}
-			return value.map(function (value) {
-				return _this.expandValue(value, props);
-			}).filter(function (i) {
-				return i;
-			});
+			return value.map(value => this.expandValue(value, props)).filter(i => i);
 		}
 		var expandedValue = this.expandValue(value, props);
 		return expandedValue ? [expandedValue] : [];
-	},
+	}
 
 	/**
   * Retrieve a value from the given options and valueKey
   * @param	{String|Number|Array}	value	- the selected value(s)
   * @param	{Object}		props	- the Select component's props (or nextProps)
   */
-	expandValue: function expandValue(value, props) {
-		var valueType = typeof value;
+	expandValue(value, props) {
+		const valueType = typeof value;
 		if (valueType !== 'string' && valueType !== 'number' && valueType !== 'boolean') return value;
-		var options = props.options;
-		var valueKey = props.valueKey;
-
+		let { options, valueKey } = props;
 		if (!options) return;
 		for (var i = 0; i < options.length; i++) {
 			if (options[i][valueKey] === value) return options[i];
 		}
-	},
+	}
 
-	setValue: function setValue(value) {
-		var _this2 = this;
-
+	setValue(value) {
 		if (this.props.autoBlur) {
 			this.blurInput();
 		}
 		if (!this.props.onChange) return;
 		if (this.props.required) {
-			var required = this.handleRequired(value, this.props.multi);
-			this.setState({ required: required });
+			const required = this.handleRequired(value, this.props.multi);
+			this.setState({ required });
 		}
 		if (this.props.simpleValue && value) {
-			value = this.props.multi ? value.map(function (i) {
-				return i[_this2.props.valueKey];
-			}).join(this.props.delimiter) : value[this.props.valueKey];
+			value = this.props.multi ? value.map(i => i[this.props.valueKey]).join(this.props.delimiter) : value[this.props.valueKey];
 		}
 		this.props.onChange(value);
-	},
+	}
 
-	selectValue: function selectValue(value) {
-		var _this3 = this;
-
+	selectValue(value) {
 		//NOTE: update value in the callback to make sure the input value is empty so that there are no styling issues (Chrome had issue otherwise)
 		this.hasScrolledToOption = false;
 		if (this.props.multi) {
+			const updatedValue = this.props.onSelectResetsInput ? '' : this.state.inputValue;
 			this.setState({
-				inputValue: '',
+				inputValue: this.handleInputValueChange(updatedValue),
 				focusedIndex: null
-			}, function () {
-				_this3.addValue(value);
+			}, () => {
+				this.addValue(value);
 			});
 		} else {
 			this.setState({
 				isOpen: false,
-				inputValue: '',
+				inputValue: this.handleInputValueChange(''),
 				isPseudoFocused: this.state.isFocused
-			}, function () {
-				_this3.setValue(value);
+			}, () => {
+				this.setValue(value);
 			});
 		}
-	},
+	}
 
-	addValue: function addValue(value) {
+	addValue(value) {
 		var valueArray = this.getValueArray(this.props.value);
-		var visibleOptions = this._visibleOptions.filter(function (val) {
-			return !val.disabled;
-		});
-		var lastValueIndex = visibleOptions.indexOf(value);
+		const visibleOptions = this._visibleOptions.filter(val => !val.disabled);
+		const lastValueIndex = visibleOptions.indexOf(value);
 		this.setValue(valueArray.concat(value));
 		if (visibleOptions.length - 1 === lastValueIndex) {
 			// the last option was selected; focus the second-last one
@@ -720,24 +562,22 @@ var Select = (0, _createReactClass2['default'])({
 			// focus the option below the selected one
 			this.focusOption(visibleOptions[lastValueIndex + 1]);
 		}
-	},
+	}
 
-	popValue: function popValue() {
+	popValue() {
 		var valueArray = this.getValueArray(this.props.value);
 		if (!valueArray.length) return;
 		if (valueArray[valueArray.length - 1].clearableValue === false) return;
-		this.setValue(valueArray.slice(0, valueArray.length - 1));
-	},
+		this.setValue(this.props.multi ? valueArray.slice(0, valueArray.length - 1) : null);
+	}
 
-	removeValue: function removeValue(value) {
+	removeValue(value) {
 		var valueArray = this.getValueArray(this.props.value);
-		this.setValue(valueArray.filter(function (i) {
-			return i !== value;
-		}));
+		this.setValue(valueArray.filter(i => i !== value));
 		this.focus();
-	},
+	}
 
-	clearValue: function clearValue(event) {
+	clearValue(event) {
 		// if the event was triggered by a mousedown and not the primary
 		// button, ignore it.
 		if (event && event.type === 'mousedown' && event.button !== 0) {
@@ -748,11 +588,11 @@ var Select = (0, _createReactClass2['default'])({
 		this.setValue(this.getResetValue());
 		this.setState({
 			isOpen: false,
-			inputValue: ''
+			inputValue: this.handleInputValueChange('')
 		}, this.focus);
-	},
+	}
 
-	getResetValue: function getResetValue() {
+	getResetValue() {
 		if (this.props.resetValue !== undefined) {
 			return this.props.resetValue;
 		} else if (this.props.multi) {
@@ -760,44 +600,40 @@ var Select = (0, _createReactClass2['default'])({
 		} else {
 			return null;
 		}
-	},
+	}
 
-	focusOption: function focusOption(option) {
+	focusOption(option) {
 		this.setState({
 			focusedOption: option
 		});
-	},
+	}
 
-	focusNextOption: function focusNextOption() {
+	focusNextOption() {
 		this.focusAdjacentOption('next');
-	},
+	}
 
-	focusPreviousOption: function focusPreviousOption() {
+	focusPreviousOption() {
 		this.focusAdjacentOption('previous');
-	},
+	}
 
-	focusPageUpOption: function focusPageUpOption() {
+	focusPageUpOption() {
 		this.focusAdjacentOption('page_up');
-	},
+	}
 
-	focusPageDownOption: function focusPageDownOption() {
+	focusPageDownOption() {
 		this.focusAdjacentOption('page_down');
-	},
+	}
 
-	focusStartOption: function focusStartOption() {
+	focusStartOption() {
 		this.focusAdjacentOption('start');
-	},
+	}
 
-	focusEndOption: function focusEndOption() {
+	focusEndOption() {
 		this.focusAdjacentOption('end');
-	},
+	}
 
-	focusAdjacentOption: function focusAdjacentOption(dir) {
-		var options = this._visibleOptions.map(function (option, index) {
-			return { option: option, index: index };
-		}).filter(function (option) {
-			return !option.option.disabled;
-		});
+	focusAdjacentOption(dir) {
+		var options = this._visibleOptions.map((option, index) => ({ option, index })).filter(option => !option.option.disabled);
 		this._scrollToFocusedOptionOnUpdate = true;
 		if (!this.state.isOpen) {
 			this.setState({
@@ -851,68 +687,66 @@ var Select = (0, _createReactClass2['default'])({
 			focusedIndex: options[focusedIndex].index,
 			focusedOption: options[focusedIndex].option
 		});
-	},
+	}
 
-	getFocusedOption: function getFocusedOption() {
+	getFocusedOption() {
 		return this._focusedOption;
-	},
+	}
 
-	getInputValue: function getInputValue() {
+	getInputValue() {
 		return this.state.inputValue;
-	},
+	}
 
-	selectFocusedOption: function selectFocusedOption() {
+	selectFocusedOption() {
 		if (this._focusedOption) {
 			return this.selectValue(this._focusedOption);
 		}
-	},
+	}
 
-	renderLoading: function renderLoading() {
+	renderLoading() {
 		if (!this.props.isLoading) return;
-		return _react2['default'].createElement(
+		return React.createElement(
 			'span',
 			{ className: 'Select-loading-zone', 'aria-hidden': 'true' },
-			_react2['default'].createElement('span', { className: 'Select-loading' })
+			React.createElement('span', { className: 'Select-loading' })
 		);
-	},
+	}
 
-	renderValue: function renderValue(valueArray, isOpen) {
-		var _this4 = this;
-
-		var renderLabel = this.props.valueRenderer || this.getOptionLabel;
-		var ValueComponent = this.props.valueComponent;
+	renderValue(valueArray, isOpen) {
+		let renderLabel = this.props.valueRenderer || this.getOptionLabel;
+		let ValueComponent = this.props.valueComponent;
 		if (!valueArray.length) {
-			return !this.state.inputValue ? _react2['default'].createElement(
+			return !this.state.inputValue ? React.createElement(
 				'div',
 				{ className: 'Select-placeholder' },
 				this.props.placeholder
 			) : null;
 		}
-		var onClick = this.props.onValueClick ? this.handleValueClick : null;
+		let onClick = this.props.onValueClick ? this.handleValueClick : null;
 		if (this.props.multi) {
-			return valueArray.map(function (value, i) {
-				return _react2['default'].createElement(
+			return valueArray.map((value, i) => {
+				return React.createElement(
 					ValueComponent,
 					{
-						id: _this4._instancePrefix + '-value-' + i,
-						instancePrefix: _this4._instancePrefix,
-						disabled: _this4.props.disabled || value.clearableValue === false,
-						key: 'value-' + i + '-' + value[_this4.props.valueKey],
+						id: this._instancePrefix + '-value-' + i,
+						instancePrefix: this._instancePrefix,
+						disabled: this.props.disabled || value.clearableValue === false,
+						key: `value-${i}-${value[this.props.valueKey]}`,
 						onClick: onClick,
-						onRemove: _this4.removeValue,
+						onRemove: this.removeValue,
 						value: value
 					},
 					renderLabel(value, i),
-					_react2['default'].createElement(
+					React.createElement(
 						'span',
 						{ className: 'Select-aria-only' },
-						' '
+						'\xA0'
 					)
 				);
 			});
 		} else if (!this.state.inputValue) {
 			if (isOpen) onClick = null;
-			return _react2['default'].createElement(
+			return React.createElement(
 				ValueComponent,
 				{
 					id: this._instancePrefix + '-value-item',
@@ -924,19 +758,19 @@ var Select = (0, _createReactClass2['default'])({
 				renderLabel(valueArray[0])
 			);
 		}
-	},
+	}
 
-	renderInput: function renderInput(valueArray, focusedOptionIndex) {
-		var _classNames,
-		    _this5 = this;
+	renderInput(valueArray, focusedOptionIndex) {
+		var className = classNames('Select-input', this.props.inputProps.className);
+		const isOpen = !!this.state.isOpen;
 
-		var className = (0, _classnames2['default'])('Select-input', this.props.inputProps.className);
-		var isOpen = !!this.state.isOpen;
-
-		var ariaOwns = (0, _classnames2['default'])((_classNames = {}, _defineProperty(_classNames, this._instancePrefix + '-list', isOpen), _defineProperty(_classNames, this._instancePrefix + '-backspace-remove-message', this.props.multi && !this.props.disabled && this.state.isFocused && !this.state.inputValue), _classNames));
+		const ariaOwns = classNames({
+			[this._instancePrefix + '-list']: isOpen,
+			[this._instancePrefix + '-backspace-remove-message']: this.props.multi && !this.props.disabled && this.state.isFocused && !this.state.inputValue
+		});
 
 		// TODO: Check how this project includes Object.assign()
-		var inputProps = _extends({}, this.props.inputProps, {
+		const inputProps = Object.assign({}, this.props.inputProps, {
 			role: 'combobox',
 			'aria-expanded': '' + isOpen,
 			'aria-owns': ariaOwns,
@@ -950,9 +784,7 @@ var Select = (0, _createReactClass2['default'])({
 			onBlur: this.handleInputBlur,
 			onChange: this.handleInputChange,
 			onFocus: this.handleInputFocus,
-			ref: function ref(_ref) {
-				return _this5.input = _ref;
-			},
+			ref: ref => this.input = ref,
 			required: this.state.required,
 			value: this.state.inputValue
 		});
@@ -962,45 +794,44 @@ var Select = (0, _createReactClass2['default'])({
 		}
 
 		if (this.props.disabled || !this.props.searchable) {
-			var _props$inputProps = this.props.inputProps;
-			var inputClassName = _props$inputProps.inputClassName;
+			const _props$inputProps = this.props.inputProps,
+			      { inputClassName } = _props$inputProps,
+			      divProps = _objectWithoutProperties(_props$inputProps, ['inputClassName']);
 
-			var divProps = _objectWithoutProperties(_props$inputProps, ['inputClassName']);
+			const ariaOwns = classNames({
+				[this._instancePrefix + '-list']: isOpen
+			});
 
-			var _ariaOwns = (0, _classnames2['default'])(_defineProperty({}, this._instancePrefix + '-list', isOpen));
-
-			return _react2['default'].createElement('div', _extends({}, divProps, {
+			return React.createElement('div', _extends({}, divProps, {
 				role: 'combobox',
 				'aria-expanded': isOpen,
-				'aria-owns': _ariaOwns,
+				'aria-owns': ariaOwns,
 				'aria-activedescendant': isOpen ? this._instancePrefix + '-option-' + focusedOptionIndex : this._instancePrefix + '-value',
 				className: className,
 				tabIndex: this.props.tabIndex || 0,
 				onBlur: this.handleInputBlur,
 				onFocus: this.handleInputFocus,
-				ref: function (ref) {
-					return _this5.input = ref;
-				},
+				ref: ref => this.input = ref,
 				'aria-readonly': '' + !!this.props.disabled,
 				style: { border: 0, width: 1, display: 'inline-block' } }));
 		}
 
 		if (this.props.autosize) {
-			return _react2['default'].createElement(_reactInputAutosize2['default'], _extends({}, inputProps, { minWidth: '5' }));
+			return React.createElement(AutosizeInput, _extends({}, inputProps, { minWidth: '5' }));
 		}
-		return _react2['default'].createElement(
+		return React.createElement(
 			'div',
 			{ className: className },
-			_react2['default'].createElement('input', inputProps)
+			React.createElement('input', inputProps)
 		);
-	},
+	}
 
-	renderClear: function renderClear() {
+	renderClear() {
 
 		if (!this.props.clearable || this.props.value === undefined || this.props.value === null || this.props.multi && !this.props.value.length || this.props.disabled || this.props.isLoading) return;
-		var clear = this.props.clearRenderer();
+		const clear = this.props.clearRenderer();
 
-		return _react2['default'].createElement(
+		return React.createElement(
 			'span',
 			{ className: 'Select-clear-zone', title: this.props.multi ? this.props.clearAllText : this.props.clearValueText,
 				'aria-label': this.props.multi ? this.props.clearAllText : this.props.clearValueText,
@@ -1011,14 +842,14 @@ var Select = (0, _createReactClass2['default'])({
 			},
 			clear
 		);
-	},
+	}
 
-	renderArrow: function renderArrow() {
-		var onMouseDown = this.handleMouseDownOnArrow;
-		var isOpen = this.state.isOpen;
-		var arrow = this.props.arrowRenderer({ onMouseDown: onMouseDown, isOpen: isOpen });
+	renderArrow() {
+		const onMouseDown = this.handleMouseDownOnArrow;
+		const isOpen = this.state.isOpen;
+		const arrow = this.props.arrowRenderer({ onMouseDown, isOpen });
 
-		return _react2['default'].createElement(
+		return React.createElement(
 			'span',
 			{
 				className: 'Select-arrow-zone',
@@ -1026,14 +857,14 @@ var Select = (0, _createReactClass2['default'])({
 			},
 			arrow
 		);
-	},
+	}
 
-	filterOptions: function filterOptions(excludeOptions) {
+	filterOptions(excludeOptions) {
 		var filterValue = this.state.inputValue;
 		var options = this.props.options || [];
 		if (this.props.filterOptions) {
 			// Maintain backwards compatibility with boolean attribute
-			var filterOptions = typeof this.props.filterOptions === 'function' ? this.props.filterOptions : _utilsDefaultFilterOptions2['default'];
+			const filterOptions = typeof this.props.filterOptions === 'function' ? this.props.filterOptions : defaultFilterOptions;
 
 			return filterOptions(options, filterValue, excludeOptions, {
 				filterOption: this.props.filterOption,
@@ -1047,18 +878,18 @@ var Select = (0, _createReactClass2['default'])({
 		} else {
 			return options;
 		}
-	},
+	}
 
-	onOptionRef: function onOptionRef(ref, isFocused) {
+	onOptionRef(ref, isFocused) {
 		if (isFocused) {
 			this.focused = ref;
 		}
-	},
+	}
 
-	renderMenu: function renderMenu(options, valueArray, focusedOption) {
+	renderMenu(options, valueArray, focusedOption) {
 		if (options && options.length) {
 			return this.props.menuRenderer({
-				focusedOption: focusedOption,
+				focusedOption,
 				focusOption: this.focusOption,
 				instancePrefix: this._instancePrefix,
 				labelKey: this.props.labelKey,
@@ -1067,14 +898,14 @@ var Select = (0, _createReactClass2['default'])({
 				optionClassName: this.props.optionClassName,
 				optionComponent: this.props.optionComponent,
 				optionRenderer: this.props.optionRenderer || this.getOptionLabel,
-				options: options,
+				options,
 				selectValue: this.selectValue,
-				valueArray: valueArray,
+				valueArray,
 				valueKey: this.props.valueKey,
 				onOptionRef: this.onOptionRef
 			});
 		} else if (this.props.noResultsText) {
-			return _react2['default'].createElement(
+			return React.createElement(
 				'div',
 				{ className: 'Select-noresults' },
 				this.props.noResultsText
@@ -1082,45 +913,37 @@ var Select = (0, _createReactClass2['default'])({
 		} else {
 			return null;
 		}
-	},
+	}
 
-	renderHiddenField: function renderHiddenField(valueArray) {
-		var _this6 = this;
-
+	renderHiddenField(valueArray) {
 		if (!this.props.name) return;
 		if (this.props.joinValues) {
-			var value = valueArray.map(function (i) {
-				return stringifyValue(i[_this6.props.valueKey]);
-			}).join(this.props.delimiter);
-			return _react2['default'].createElement('input', {
+			let value = valueArray.map(i => stringifyValue(i[this.props.valueKey])).join(this.props.delimiter);
+			return React.createElement('input', {
 				type: 'hidden',
-				ref: function (ref) {
-					return _this6.value = ref;
-				},
+				ref: ref => this.value = ref,
 				name: this.props.name,
 				value: value,
 				disabled: this.props.disabled });
 		}
-		return valueArray.map(function (item, index) {
-			return _react2['default'].createElement('input', { key: 'hidden.' + index,
-				type: 'hidden',
-				ref: 'value' + index,
-				name: _this6.props.name,
-				value: stringifyValue(item[_this6.props.valueKey]),
-				disabled: _this6.props.disabled });
-		});
-	},
+		return valueArray.map((item, index) => React.createElement('input', { key: 'hidden.' + index,
+			type: 'hidden',
+			ref: 'value' + index,
+			name: this.props.name,
+			value: stringifyValue(item[this.props.valueKey]),
+			disabled: this.props.disabled }));
+	}
 
-	getFocusableOptionIndex: function getFocusableOptionIndex(selectedOption) {
+	getFocusableOptionIndex(selectedOption) {
 		var options = this._visibleOptions;
 		if (!options.length) return null;
 
-		var valueKey = this.props.valueKey;
-		var focusedOption = this.state.focusedOption || selectedOption;
+		const valueKey = this.props.valueKey;
+		let focusedOption = this.state.focusedOption || selectedOption;
 		if (focusedOption && !focusedOption.disabled) {
-			var focusedOptionIndex = -1;
-			options.some(function (option, index) {
-				var isOptionEqual = option[valueKey] === focusedOption[valueKey];
+			let focusedOptionIndex = -1;
+			options.some((option, index) => {
+				const isOptionEqual = option[valueKey] === focusedOption[valueKey];
 				if (isOptionEqual) {
 					focusedOptionIndex = index;
 				}
@@ -1135,50 +958,42 @@ var Select = (0, _createReactClass2['default'])({
 			if (!options[i].disabled) return i;
 		}
 		return null;
-	},
+	}
 
-	renderOuter: function renderOuter(options, valueArray, focusedOption) {
-		var _this7 = this;
-
-		var menu = this.renderMenu(options, valueArray, focusedOption);
+	renderOuter(options, valueArray, focusedOption) {
+		let menu = this.renderMenu(options, valueArray, focusedOption);
 		if (!menu) {
 			return null;
 		}
 
-		return _react2['default'].createElement(
+		return React.createElement(
 			'div',
-			{ ref: function (ref) {
-					return _this7.menuContainer = ref;
-				}, className: 'Select-menu-outer', style: this.props.menuContainerStyle },
-			_react2['default'].createElement(
+			{ ref: ref => this.menuContainer = ref, className: 'Select-menu-outer', style: this.props.menuContainerStyle },
+			React.createElement(
 				'div',
-				{ ref: function (ref) {
-						return _this7.menu = ref;
-					}, role: 'listbox', className: 'Select-menu', id: this._instancePrefix + '-list',
+				{ ref: ref => this.menu = ref, role: 'listbox', tabIndex: -1, className: 'Select-menu', id: this._instancePrefix + '-list',
 					style: this.props.menuStyle,
 					onScroll: this.handleMenuScroll,
 					onMouseDown: this.handleMouseDownOnMenu },
 				menu
 			)
 		);
-	},
+	}
 
-	render: function render() {
-		var _this8 = this;
-
-		var valueArray = this.getValueArray(this.props.value);
-		var options = this._visibleOptions = this.filterOptions(this.props.multi ? this.getValueArray(this.props.value) : null);
-		var isOpen = this.state.isOpen;
+	render() {
+		let valueArray = this.getValueArray(this.props.value);
+		let options = this._visibleOptions = this.filterOptions(this.props.multi ? this.getValueArray(this.props.value) : null);
+		let isOpen = this.state.isOpen;
 		if (this.props.multi && !options.length && valueArray.length && !this.state.inputValue) isOpen = false;
-		var focusedOptionIndex = this.getFocusableOptionIndex(valueArray[0]);
+		const focusedOptionIndex = this.getFocusableOptionIndex(valueArray[0]);
 
-		var focusedOption = null;
+		let focusedOption = null;
 		if (focusedOptionIndex !== null) {
 			focusedOption = this._focusedOption = options[focusedOptionIndex];
 		} else {
 			focusedOption = this._focusedOption = null;
 		}
-		var className = (0, _classnames2['default'])('Select', this.props.className, {
+		let className = classNames('Select', this.props.className, {
 			'Select--multi': this.props.multi,
 			'Select--single': !this.props.multi,
 			'is-clearable': this.props.clearable,
@@ -1191,28 +1006,24 @@ var Select = (0, _createReactClass2['default'])({
 			'has-value': valueArray.length
 		});
 
-		var removeMessage = null;
+		let removeMessage = null;
 		if (this.props.multi && !this.props.disabled && valueArray.length && !this.state.inputValue && this.state.isFocused && this.props.backspaceRemoves) {
-			removeMessage = _react2['default'].createElement(
+			removeMessage = React.createElement(
 				'span',
 				{ id: this._instancePrefix + '-backspace-remove-message', className: 'Select-aria-only', 'aria-live': 'assertive' },
 				this.props.backspaceToRemoveMessage.replace('{label}', valueArray[valueArray.length - 1][this.props.labelKey])
 			);
 		}
 
-		return _react2['default'].createElement(
+		return React.createElement(
 			'div',
-			{ ref: function (ref) {
-					return _this8.wrapper = ref;
-				},
+			{ ref: ref => this.wrapper = ref,
 				className: className,
 				style: this.props.wrapperStyle },
 			this.renderHiddenField(valueArray),
-			_react2['default'].createElement(
+			React.createElement(
 				'div',
-				{ ref: function (ref) {
-						return _this8.control = ref;
-					},
+				{ ref: ref => this.control = ref,
 					className: 'Select-control',
 					style: this.props.style,
 					onKeyDown: this.handleKeyDown,
@@ -1221,7 +1032,7 @@ var Select = (0, _createReactClass2['default'])({
 					onTouchStart: this.handleTouchStart,
 					onTouchMove: this.handleTouchMove
 				},
-				_react2['default'].createElement(
+				React.createElement(
 					'span',
 					{ className: 'Select-multi-value-wrapper', id: this._instancePrefix + '-value' },
 					this.renderValue(valueArray, isOpen),
@@ -1235,8 +1046,122 @@ var Select = (0, _createReactClass2['default'])({
 			isOpen ? this.renderOuter(options, !this.props.multi ? valueArray : null, focusedOption) : null
 		);
 	}
+};
 
-});
+Select.propTypes = {
+	'aria-describedby': PropTypes.string, // HTML ID(s) of element(s) that should be used to describe this input (for assistive tech)
+	'aria-label': PropTypes.string, // Aria label (for assistive tech)
+	'aria-labelledby': PropTypes.string, // HTML ID of an element that should be used as the label (for assistive tech)
+	addLabelText: PropTypes.string, // placeholder displayed when you want to add a label on a multi-value input
+	arrowRenderer: PropTypes.func, // Create drop-down caret element
+	autoBlur: PropTypes.bool, // automatically blur the component when an option is selected
+	autofocus: PropTypes.bool, // autofocus the component on mount
+	autosize: PropTypes.bool, // whether to enable autosizing or not
+	backspaceRemoves: PropTypes.bool, // whether backspace removes an item if there is no text input
+	backspaceToRemoveMessage: PropTypes.string, // Message to use for screenreaders to press backspace to remove the current item - {label} is replaced with the item label
+	className: PropTypes.string, // className for the outer element
+	clearAllText: stringOrNode, // title for the "clear" control when multi: true
+	clearRenderer: PropTypes.func, // create clearable x element
+	clearValueText: stringOrNode, // title for the "clear" control
+	clearable: PropTypes.bool, // should it be possible to reset value
+	deleteRemoves: PropTypes.bool, // whether backspace removes an item if there is no text input
+	delimiter: PropTypes.string, // delimiter to use to join multiple values for the hidden field value
+	disabled: PropTypes.bool, // whether the Select is disabled or not
+	escapeClearsValue: PropTypes.bool, // whether escape clears the value when the menu is closed
+	filterOption: PropTypes.func, // method to filter a single option (option, filterString)
+	filterOptions: PropTypes.any, // boolean to enable default filtering or function to filter the options array ([options], filterString, [values])
+	ignoreAccents: PropTypes.bool, // whether to strip diacritics when filtering
+	ignoreCase: PropTypes.bool, // whether to perform case-insensitive filtering
+	inputProps: PropTypes.object, // custom attributes for the Input
+	inputRenderer: PropTypes.func, // returns a custom input component
+	instanceId: PropTypes.string, // set the components instanceId
+	isLoading: PropTypes.bool, // whether the Select is loading externally or not (such as options being loaded)
+	joinValues: PropTypes.bool, // joins multiple values into a single form field with the delimiter (legacy mode)
+	labelKey: PropTypes.string, // path of the label value in option objects
+	matchPos: PropTypes.string, // (any|start) match the start or entire string when filtering
+	matchProp: PropTypes.string, // (any|label|value) which option property to filter on
+	menuBuffer: PropTypes.number, // optional buffer (in px) between the bottom of the viewport and the bottom of the menu
+	menuContainerStyle: PropTypes.object, // optional style to apply to the menu container
+	menuRenderer: PropTypes.func, // renders a custom menu with options
+	menuStyle: PropTypes.object, // optional style to apply to the menu
+	multi: PropTypes.bool, // multi-value input
+	name: PropTypes.string, // generates a hidden <input /> tag with this field name for html forms
+	noResultsText: stringOrNode, // placeholder displayed when there are no matching search results
+	onBlur: PropTypes.func, // onBlur handler: function (event) {}
+	onBlurResetsInput: PropTypes.bool, // whether input is cleared on blur
+	onChange: PropTypes.func, // onChange handler: function (newValue) {}
+	onClose: PropTypes.func, // fires when the menu is closed
+	onCloseResetsInput: PropTypes.bool, // whether input is cleared when menu is closed through the arrow
+	onFocus: PropTypes.func, // onFocus handler: function (event) {}
+	onInputChange: PropTypes.func, // onInputChange handler: function (inputValue) {}
+	onInputKeyDown: PropTypes.func, // input keyDown handler: function (event) {}
+	onMenuScrollToBottom: PropTypes.func, // fires when the menu is scrolled to the bottom; can be used to paginate options
+	onOpen: PropTypes.func, // fires when the menu is opened
+	onSelectResetsInput: PropTypes.bool, // whether input is cleared on select (works only for multiselect)
+	onValueClick: PropTypes.func, // onClick handler for value labels: function (value, event) {}
+	openAfterFocus: PropTypes.bool, // boolean to enable opening dropdown when focused
+	openOnFocus: PropTypes.bool, // always open options menu on focus
+	optionClassName: PropTypes.string, // additional class(es) to apply to the <Option /> elements
+	optionComponent: PropTypes.func, // option component to render in dropdown
+	optionRenderer: PropTypes.func, // optionRenderer: function (option) {}
+	options: PropTypes.array, // array of options
+	pageSize: PropTypes.number, // number of entries to page when using page up/down keys
+	placeholder: stringOrNode, // field placeholder, displayed when there's no value
+	required: PropTypes.bool, // applies HTML5 required attribute when needed
+	resetValue: PropTypes.any, // value to use when you clear the control
+	scrollMenuIntoView: PropTypes.bool, // boolean to enable the viewport to shift so that the full menu fully visible when engaged
+	searchable: PropTypes.bool, // whether to enable searching feature or not
+	simpleValue: PropTypes.bool, // pass the value to onChange as a simple value (legacy pre 1.0 mode), defaults to false
+	style: PropTypes.object, // optional style to apply to the control
+	tabIndex: PropTypes.string, // optional tab index of the control
+	tabSelectsValue: PropTypes.bool, // whether to treat tabbing out while focused to be value selection
+	value: PropTypes.any, // initial field value
+	valueComponent: PropTypes.func, // value component to render
+	valueKey: PropTypes.string, // path of the label value in option objects
+	valueRenderer: PropTypes.func, // valueRenderer: function (option) {}
+	wrapperStyle: PropTypes.object // optional style to apply to the component wrapper
+};
 
-exports['default'] = Select;
-module.exports = exports['default'];
+Select.defaultProps = {
+	addLabelText: 'Add "{label}"?',
+	arrowRenderer: defaultArrowRenderer,
+	autosize: true,
+	backspaceRemoves: true,
+	backspaceToRemoveMessage: 'Press backspace to remove {label}',
+	clearable: true,
+	clearAllText: 'Clear all',
+	clearRenderer: defaultClearRenderer,
+	clearValueText: 'Clear value',
+	deleteRemoves: true,
+	delimiter: ',',
+	disabled: false,
+	escapeClearsValue: true,
+	filterOptions: defaultFilterOptions,
+	ignoreAccents: true,
+	ignoreCase: true,
+	inputProps: {},
+	isLoading: false,
+	joinValues: false,
+	labelKey: 'label',
+	matchPos: 'any',
+	matchProp: 'any',
+	menuBuffer: 0,
+	menuRenderer: defaultMenuRenderer,
+	multi: false,
+	noResultsText: 'No results found',
+	onBlurResetsInput: true,
+	onSelectResetsInput: true,
+	onCloseResetsInput: true,
+	optionComponent: Option,
+	pageSize: 5,
+	placeholder: 'Select...',
+	required: false,
+	scrollMenuIntoView: true,
+	searchable: true,
+	simpleValue: false,
+	tabSelectsValue: true,
+	valueComponent: Value,
+	valueKey: 'value'
+};
+
+export default Select;
