@@ -1,37 +1,20 @@
-'use strict';
+import React from 'react';
+import PropTypes from 'prop-types';
+import classNames from 'classnames';
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+class Value extends React.Component {
 
-var _react = require('react');
+	constructor(props) {
+		super(props);
 
-var _react2 = _interopRequireDefault(_react);
+		this.handleMouseDown = this.handleMouseDown.bind(this);
+		this.onRemove = this.onRemove.bind(this);
+		this.handleTouchEndRemove = this.handleTouchEndRemove.bind(this);
+		this.handleTouchMove = this.handleTouchMove.bind(this);
+		this.handleTouchStart = this.handleTouchStart.bind(this);
+	}
 
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
-
-var _propTypes = require('prop-types');
-
-var _propTypes2 = _interopRequireDefault(_propTypes);
-
-var _classnames = require('classnames');
-
-var _classnames2 = _interopRequireDefault(_classnames);
-
-var Value = (0, _createReactClass2['default'])({
-
-	displayName: 'Value',
-
-	propTypes: {
-		children: _propTypes2['default'].node,
-		disabled: _propTypes2['default'].bool, // disabled prop passed to ReactSelect
-		id: _propTypes2['default'].string, // Unique id for the value - used for aria
-		onClick: _propTypes2['default'].func, // method to handle click on value label
-		onRemove: _propTypes2['default'].func, // method to handle removal of the value
-		value: _propTypes2['default'].object.isRequired },
-
-	// the option object for this value
-	handleMouseDown: function handleMouseDown(event) {
+	handleMouseDown(event) {
 		if (event.type === 'mousedown' && event.button !== 0) {
 			return;
 		}
@@ -43,36 +26,36 @@ var Value = (0, _createReactClass2['default'])({
 		if (this.props.value.href) {
 			event.stopPropagation();
 		}
-	},
+	}
 
-	onRemove: function onRemove(event) {
+	onRemove(event) {
 		event.preventDefault();
 		event.stopPropagation();
 		this.props.onRemove(this.props.value);
-	},
+	}
 
-	handleTouchEndRemove: function handleTouchEndRemove(event) {
+	handleTouchEndRemove(event) {
 		// Check if the view is being dragged, In this case
 		// we don't want to fire the click event (because the user only wants to scroll)
 		if (this.dragging) return;
 
 		// Fire the mouse events
 		this.onRemove(event);
-	},
+	}
 
-	handleTouchMove: function handleTouchMove(event) {
+	handleTouchMove(event) {
 		// Set a flag that the view is being dragged
 		this.dragging = true;
-	},
+	}
 
-	handleTouchStart: function handleTouchStart(event) {
+	handleTouchStart(event) {
 		// Set a flag that the view is not being dragged
 		this.dragging = false;
-	},
+	}
 
-	renderRemoveIcon: function renderRemoveIcon() {
+	renderRemoveIcon() {
 		if (this.props.disabled || !this.props.onRemove) return;
-		return _react2['default'].createElement(
+		return React.createElement(
 			'span',
 			{ className: 'Select-value-icon',
 				'aria-hidden': 'true',
@@ -80,27 +63,27 @@ var Value = (0, _createReactClass2['default'])({
 				onTouchEnd: this.handleTouchEndRemove,
 				onTouchStart: this.handleTouchStart,
 				onTouchMove: this.handleTouchMove },
-			'×'
+			'\xD7'
 		);
-	},
+	}
 
-	renderLabel: function renderLabel() {
-		var className = 'Select-value-label';
-		return this.props.onClick || this.props.value.href ? _react2['default'].createElement(
+	renderLabel() {
+		let className = 'Select-value-label';
+		return this.props.onClick || this.props.value.href ? React.createElement(
 			'a',
 			{ className: className, href: this.props.value.href, target: this.props.value.target, onMouseDown: this.handleMouseDown, onTouchEnd: this.handleMouseDown },
 			this.props.children
-		) : _react2['default'].createElement(
+		) : React.createElement(
 			'span',
 			{ className: className, role: 'option', 'aria-selected': 'true', id: this.props.id },
 			this.props.children
 		);
-	},
+	}
 
-	render: function render() {
-		return _react2['default'].createElement(
+	render() {
+		return React.createElement(
 			'div',
-			{ className: (0, _classnames2['default'])('Select-value', this.props.value.className),
+			{ className: classNames('Select-value', this.props.value.className),
 				style: this.props.value.style,
 				title: this.props.value.title
 			},
@@ -108,7 +91,15 @@ var Value = (0, _createReactClass2['default'])({
 			this.renderLabel()
 		);
 	}
+};
 
-});
+Value.propTypes = {
+	children: PropTypes.node,
+	disabled: PropTypes.bool, // disabled prop passed to ReactSelect
+	id: PropTypes.string, // Unique id for the value - used for aria
+	onClick: PropTypes.func, // method to handle click on value label
+	onRemove: PropTypes.func, // method to handle removal of the value
+	value: PropTypes.object.isRequired // the option object for this value
+};
 
-module.exports = Value;
+export default Value;
