@@ -1,35 +1,25 @@
-'use strict';
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-var _stripDiacritics = require('./stripDiacritics');
-
-var _stripDiacritics2 = _interopRequireDefault(_stripDiacritics);
+import stripDiacritics from './stripDiacritics';
 
 function filterOptions(options, filterValue, excludeOptions, props) {
-	var _this = this;
-
 	if (props.ignoreAccents) {
-		filterValue = (0, _stripDiacritics2['default'])(filterValue);
+		filterValue = stripDiacritics(filterValue);
 	}
 
 	if (props.ignoreCase) {
 		filterValue = filterValue.toLowerCase();
 	}
 
-	if (excludeOptions) excludeOptions = excludeOptions.map(function (i) {
-		return i[props.valueKey];
-	});
+	if (excludeOptions) excludeOptions = excludeOptions.map(i => i[props.valueKey]);
 
-	return options.filter(function (option) {
+	return options.filter(option => {
 		if (excludeOptions && excludeOptions.indexOf(option[props.valueKey]) > -1) return false;
-		if (props.filterOption) return props.filterOption.call(_this, option, filterValue);
+		if (props.filterOption) return props.filterOption.call(this, option, filterValue);
 		if (!filterValue) return true;
 		var valueTest = String(option[props.valueKey]);
 		var labelTest = String(option[props.labelKey]);
 		if (props.ignoreAccents) {
-			if (props.matchProp !== 'label') valueTest = (0, _stripDiacritics2['default'])(valueTest);
-			if (props.matchProp !== 'value') labelTest = (0, _stripDiacritics2['default'])(labelTest);
+			if (props.matchProp !== 'label') valueTest = stripDiacritics(valueTest);
+			if (props.matchProp !== 'value') labelTest = stripDiacritics(labelTest);
 		}
 		if (props.ignoreCase) {
 			if (props.matchProp !== 'label') valueTest = valueTest.toLowerCase();
@@ -39,4 +29,4 @@ function filterOptions(options, filterValue, excludeOptions, props) {
 	});
 }
 
-module.exports = filterOptions;
+export default filterOptions;
