@@ -1,42 +1,33 @@
-'use strict';
+import classNames from 'classnames';
+import React from 'react';
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+function menuRenderer({
+	focusedOption,
+	instancePrefix,
+	labelKey,
+	onFocus,
+	onSelect,
+	optionClassName,
+	optionComponent,
+	optionRenderer,
+	options,
+	valueArray,
+	valueKey,
+	onOptionRef
+}) {
+	let Option = optionComponent;
 
-var _classnames = require('classnames');
-
-var _classnames2 = _interopRequireDefault(_classnames);
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-function menuRenderer(_ref) {
-	var focusedOption = _ref.focusedOption;
-	var instancePrefix = _ref.instancePrefix;
-	var labelKey = _ref.labelKey;
-	var onFocus = _ref.onFocus;
-	var onSelect = _ref.onSelect;
-	var optionClassName = _ref.optionClassName;
-	var optionComponent = _ref.optionComponent;
-	var optionRenderer = _ref.optionRenderer;
-	var options = _ref.options;
-	var valueArray = _ref.valueArray;
-	var valueKey = _ref.valueKey;
-	var onOptionRef = _ref.onOptionRef;
-
-	var Option = optionComponent;
-
-	return options.map(function (option, i) {
-		var isSelected = valueArray && valueArray.indexOf(option) > -1;
-		var isFocused = option === focusedOption;
-		var optionClass = (0, _classnames2['default'])(optionClassName, {
+	return options.map((option, i) => {
+		let isSelected = valueArray && valueArray.indexOf(option) > -1;
+		let isFocused = option === focusedOption;
+		let optionClass = classNames(optionClassName, {
 			'Select-option': true,
 			'is-selected': isSelected,
 			'is-focused': isFocused,
 			'is-disabled': option.disabled
 		});
 
-		return _react2['default'].createElement(
+		return React.createElement(
 			Option,
 			{
 				className: optionClass,
@@ -44,12 +35,12 @@ function menuRenderer(_ref) {
 				isDisabled: option.disabled,
 				isFocused: isFocused,
 				isSelected: isSelected,
-				key: 'option-' + i + '-' + option[valueKey],
+				key: `option-${i}-${option[valueKey]}`,
 				onFocus: onFocus,
 				onSelect: onSelect,
 				option: option,
 				optionIndex: i,
-				ref: function (ref) {
+				ref: ref => {
 					onOptionRef(ref, isFocused);
 				}
 			},
@@ -58,4 +49,4 @@ function menuRenderer(_ref) {
 	});
 }
 
-module.exports = menuRenderer;
+export default menuRenderer;
