@@ -1,11 +1,9 @@
 /*!
-  Copyright (c) 2016 Jed Watson.
+  Copyright (c) 2017 Jed Watson.
   Licensed under the MIT License (MIT), see
   http://jedwatson.github.io/react-select
 */
-
 import React from 'react';
-import createClass from 'create-react-class';
 import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import AutosizeInput from 'react-input-autosize';
@@ -16,9 +14,6 @@ import defaultFilterOptions from './utils/defaultFilterOptions';
 import defaultMenuRenderer from './utils/defaultMenuRenderer';
 import defaultClearRenderer from './utils/defaultClearRenderer';
 
-import Async from './Async';
-import AsyncCreatable from './AsyncCreatable';
-import Creatable from './Creatable';
 import Option from './Option';
 import Value from './Value';
 
@@ -42,137 +37,42 @@ const stringOrNode = PropTypes.oneOfType([
 
 let instanceId = 1;
 
-const Select = createClass({
+class Select extends React.Component {
 
-	displayName: 'Select',
+	constructor(props) {
+		super(props);
 
-	propTypes: {
-		addLabelText: PropTypes.string,       // placeholder displayed when you want to add a label on a multi-value input
-		'aria-describedby': PropTypes.string,	// HTML ID(s) of element(s) that should be used to describe this input (for assistive tech)
-		'aria-label': PropTypes.string,       // Aria label (for assistive tech)
-		'aria-labelledby': PropTypes.string,	// HTML ID of an element that should be used as the label (for assistive tech)
-		arrowRenderer: PropTypes.func,				// Create drop-down caret element
-		autoBlur: PropTypes.bool,             // automatically blur the component when an option is selected
-		autofocus: PropTypes.bool,            // autofocus the component on mount
-		autosize: PropTypes.bool,             // whether to enable autosizing or not
-		backspaceRemoves: PropTypes.bool,     // whether backspace removes an item if there is no text input
-		backspaceToRemoveMessage: PropTypes.string,  // Message to use for screenreaders to press backspace to remove the current item - {label} is replaced with the item label
-		className: PropTypes.string,          // className for the outer element
-		clearAllText: stringOrNode,                 // title for the "clear" control when multi: true
-		clearRenderer: PropTypes.func,        // create clearable x element
-		clearValueText: stringOrNode,               // title for the "clear" control
-		clearable: PropTypes.bool,            // should it be possible to reset value
-		deleteRemoves: PropTypes.bool,        // whether backspace removes an item if there is no text input
-		delimiter: PropTypes.string,          // delimiter to use to join multiple values for the hidden field value
-		disabled: PropTypes.bool,             // whether the Select is disabled or not
-		escapeClearsValue: PropTypes.bool,    // whether escape clears the value when the menu is closed
-		filterOption: PropTypes.func,         // method to filter a single option (option, filterString)
-		filterOptions: PropTypes.any,         // boolean to enable default filtering or function to filter the options array ([options], filterString, [values])
-		ignoreAccents: PropTypes.bool,        // whether to strip diacritics when filtering
-		ignoreCase: PropTypes.bool,           // whether to perform case-insensitive filtering
-		inputProps: PropTypes.object,         // custom attributes for the Input
-		inputRenderer: PropTypes.func,        // returns a custom input component
-		instanceId: PropTypes.string,         // set the components instanceId
-		isLoading: PropTypes.bool,            // whether the Select is loading externally or not (such as options being loaded)
-		joinValues: PropTypes.bool,           // joins multiple values into a single form field with the delimiter (legacy mode)
-		labelKey: PropTypes.string,           // path of the label value in option objects
-		matchPos: PropTypes.string,           // (any|start) match the start or entire string when filtering
-		matchProp: PropTypes.string,          // (any|label|value) which option property to filter on
-		menuBuffer: PropTypes.number,         // optional buffer (in px) between the bottom of the viewport and the bottom of the menu
-		menuContainerStyle: PropTypes.object, // optional style to apply to the menu container
-		menuRenderer: PropTypes.func,         // renders a custom menu with options
-		menuStyle: PropTypes.object,          // optional style to apply to the menu
-		multi: PropTypes.bool,                // multi-value input
-		name: PropTypes.string,               // generates a hidden <input /> tag with this field name for html forms
-		noResultsText: stringOrNode,                // placeholder displayed when there are no matching search results
-		onBlur: PropTypes.func,               // onBlur handler: function (event) {}
-		onBlurResetsInput: PropTypes.bool,    // whether input is cleared on blur
-		onChange: PropTypes.func,             // onChange handler: function (newValue) {}
-		onClose: PropTypes.func,              // fires when the menu is closed
-		onCloseResetsInput: PropTypes.bool,		// whether input is cleared when menu is closed through the arrow
-		onFocus: PropTypes.func,              // onFocus handler: function (event) {}
-		onInputChange: PropTypes.func,        // onInputChange handler: function (inputValue) {}
-		onInputKeyDown: PropTypes.func,       // input keyDown handler: function (event) {}
-		onMenuScrollToBottom: PropTypes.func, // fires when the menu is scrolled to the bottom; can be used to paginate options
-		onOpen: PropTypes.func,               // fires when the menu is opened
-		onValueClick: PropTypes.func,         // onClick handler for value labels: function (value, event) {}
-		openAfterFocus: PropTypes.bool,       // boolean to enable opening dropdown when focused
-		openOnFocus: PropTypes.bool,          // always open options menu on focus
-		optionClassName: PropTypes.string,    // additional class(es) to apply to the <Option /> elements
-		optionComponent: PropTypes.func,      // option component to render in dropdown
-		optionRenderer: PropTypes.func,       // optionRenderer: function (option) {}
-		options: PropTypes.array,             // array of options
-		pageSize: PropTypes.number,           // number of entries to page when using page up/down keys
-		placeholder: stringOrNode,                  // field placeholder, displayed when there's no value
-		required: PropTypes.bool,             // applies HTML5 required attribute when needed
-		resetValue: PropTypes.any,            // value to use when you clear the control
-		scrollMenuIntoView: PropTypes.bool,   // boolean to enable the viewport to shift so that the full menu fully visible when engaged
-		searchable: PropTypes.bool,           // whether to enable searching feature or not
-		simpleValue: PropTypes.bool,          // pass the value to onChange as a simple value (legacy pre 1.0 mode), defaults to false
-		style: PropTypes.object,              // optional style to apply to the control
-		tabIndex: PropTypes.string,           // optional tab index of the control
-		tabSelectsValue: PropTypes.bool,      // whether to treat tabbing out while focused to be value selection
-		value: PropTypes.any,                 // initial field value
-		valueComponent: PropTypes.func,       // value component to render
-		valueKey: PropTypes.string,           // path of the label value in option objects
-		valueRenderer: PropTypes.func,        // valueRenderer: function (option) {}
-		wrapperStyle: PropTypes.object,       // optional style to apply to the component wrapper
-	},
+		this.handleTouchOutside = this.handleTouchOutside.bind(this);
+		this.handleTouchMove = this.handleTouchMove.bind(this);
+		this.handleTouchStart = this.handleTouchStart.bind(this);
+		this.handleTouchEnd = this.handleTouchEnd.bind(this);
+		this.handleTouchEndClearValue = this.handleTouchEndClearValue.bind(this);
+		this.handleMouseDown = this.handleMouseDown.bind(this);
+		this.handleMouseDownOnArrow = this.handleMouseDownOnArrow.bind(this);
+		this.handleMouseDownOnMenu = this.handleMouseDownOnMenu.bind(this);
+		this.handleInputFocus = this.handleInputFocus.bind(this);
+		this.handleInputBlur = this.handleInputBlur.bind(this);
+		this.handleInputChange = this.handleInputChange.bind(this);
+		this.handleInputValueChange = this.handleInputValueChange.bind(this);
+		this.handleKeyDown = this.handleKeyDown.bind(this);
+		this.handleValueClick = this.handleValueClick.bind(this);
+		this.handleMenuScroll = this.handleMenuScroll.bind(this);
+		this.handleRequired = this.handleRequired.bind(this);
+		this.getOptionLabel = this.getOptionLabel.bind(this);
+		this.onOptionRef = this.onOptionRef.bind(this);
+		this.clearValue = this.clearValue.bind(this);
+		this.removeValue = this.removeValue.bind(this);
+		this.selectValue = this.selectValue.bind(this);
+		this.focusOption = this.focusOption.bind(this);
 
-	statics: { Async, AsyncCreatable, Creatable },
-
-	getDefaultProps () {
-		return {
-			addLabelText: 'Add "{label}"?',
-			arrowRenderer: defaultArrowRenderer,
-			autosize: true,
-			backspaceRemoves: true,
-			backspaceToRemoveMessage: 'Press backspace to remove {label}',
-			clearable: true,
-			clearAllText: 'Clear all',
-			clearRenderer: defaultClearRenderer,
-			clearValueText: 'Clear value',
-			deleteRemoves: true,
-			delimiter: ',',
-			disabled: false,
-			escapeClearsValue: true,
-			filterOptions: defaultFilterOptions,
-			ignoreAccents: true,
-			ignoreCase: true,
-			inputProps: {},
-			isLoading: false,
-			joinValues: false,
-			labelKey: 'label',
-			matchPos: 'any',
-			matchProp: 'any',
-			menuBuffer: 0,
-			menuRenderer: defaultMenuRenderer,
-			multi: false,
-			noResultsText: 'No results found',
-			onBlurResetsInput: true,
-			onCloseResetsInput: true,
-			optionComponent: Option,
-			pageSize: 5,
-			placeholder: 'Select...',
-			required: false,
-			scrollMenuIntoView: true,
-			searchable: true,
-			simpleValue: false,
-			tabSelectsValue: true,
-			valueComponent: Value,
-			valueKey: 'value',
-		};
-	},
-
-	getInitialState () {
-		return {
+		this.state = {
 			inputValue: '',
 			isFocused: false,
 			isOpen: false,
 			isPseudoFocused: false,
 			required: false,
 		};
-	},
+	}
 
 	componentWillMount () {
 		this._instancePrefix = 'react-select-' + (this.props.instanceId || ++instanceId) + '-';
@@ -183,13 +83,13 @@ const Select = createClass({
 				required: this.handleRequired(valueArray[0], this.props.multi),
 			});
 		}
-	},
+	}
 
 	componentDidMount () {
 		if (this.props.autofocus) {
 			this.focus();
 		}
-	},
+	}
 
 	componentWillReceiveProps (nextProps) {
 		const valueArray = this.getValueArray(nextProps.value, nextProps);
@@ -198,8 +98,11 @@ const Select = createClass({
 			this.setState({
 				required: this.handleRequired(valueArray[0], nextProps.multi),
 			});
+		} else if (this.props.required) {
+			// Used to be required but it's not any more
+			this.setState({ required: false });
 		}
-	},
+	}
 
 	componentWillUpdate (nextProps, nextState) {
 		if (nextState.isOpen !== this.state.isOpen) {
@@ -207,7 +110,7 @@ const Select = createClass({
 			const handler = nextState.isOpen ? nextProps.onOpen : nextProps.onClose;
 			handler && handler();
 		}
-	},
+	}
 
 	componentDidUpdate (prevProps, prevState) {
 		// focus to the selected option
@@ -240,7 +143,7 @@ const Select = createClass({
 			this.setState({ isFocused: false }); // eslint-disable-line react/no-did-update-set-state
 			this.closeMenu();
 		}
-	},
+	}
 
 	componentWillUnmount () {
 		if (!document.removeEventListener && document.detachEvent) {
@@ -248,7 +151,7 @@ const Select = createClass({
 		} else {
 			document.removeEventListener('touchstart', this.handleTouchOutside);
 		}
-	},
+	}
 
 	toggleTouchOutsideEvent (enabled) {
 		if (enabled) {
@@ -264,34 +167,34 @@ const Select = createClass({
 				document.removeEventListener('touchstart', this.handleTouchOutside);
 			}
 		}
-	},
+	}
 
 	handleTouchOutside (event) {
 		// handle touch outside on ios to dismiss menu
 		if (this.wrapper && !this.wrapper.contains(event.target)) {
 			this.closeMenu();
 		}
-	},
+	}
 
 	focus () {
 		if (!this.input) return;
 		this.input.focus();
-	},
+	}
 
 	blurInput () {
 		if (!this.input) return;
 		this.input.blur();
-	},
+	}
 
 	handleTouchMove (event) {
 		// Set a flag that the view is being dragged
 		this.dragging = true;
-	},
+	}
 
 	handleTouchStart (event) {
 		// Set a flag that the view is not being dragged
 		this.dragging = false;
-	},
+	}
 
 	handleTouchEnd (event) {
 		// Check if the view is being dragged, In this case
@@ -300,7 +203,7 @@ const Select = createClass({
 
 		// Fire the mouse events
 		this.handleMouseDown(event);
-	},
+	}
 
 	handleTouchEndClearValue (event) {
 		// Check if the view is being dragged, In this case
@@ -309,7 +212,7 @@ const Select = createClass({
 
 		// Clear the value
 		this.clearValue(event);
-	},
+	}
 
 	handleMouseDown (event) {
 		// if the event was triggered by a mousedown and not the primary
@@ -328,6 +231,7 @@ const Select = createClass({
 
 		// for the non-searchable select, toggle the menu
 		if (!this.props.searchable) {
+			// TODO: This code means that if a select is searchable, onClick the options menu will not appear, only on subsequent click will it open.
 			this.focus();
 			return this.setState({
 				isOpen: !this.state.isOpen,
@@ -356,10 +260,10 @@ const Select = createClass({
 			});
 		} else {
 			// otherwise, focus the input and open the menu
-			this._openAfterFocus = true;
+			this._openAfterFocus = this.props.openAfterFocus;
 			this.focus();
 		}
-	},
+	}
 
 	handleMouseDownOnArrow (event) {
 		// if the event was triggered by a mousedown and not the primary
@@ -376,7 +280,7 @@ const Select = createClass({
 		event.preventDefault();
 		// close the menu
 		this.closeMenu();
-	},
+	}
 
 	handleMouseDownOnMenu (event) {
 		// if the event was triggered by a mousedown and not the primary
@@ -389,24 +293,23 @@ const Select = createClass({
 
 		this._openAfterFocus = true;
 		this.focus();
-	},
+	}
 
 	closeMenu () {
 		if(this.props.onCloseResetsInput) {
 			this.setState({
 				isOpen: false,
 				isPseudoFocused: this.state.isFocused && !this.props.multi,
-				inputValue: ''
+				inputValue: this.handleInputValueChange('')
 			});
 		} else {
 			this.setState({
 				isOpen: false,
-				isPseudoFocused: this.state.isFocused && !this.props.multi,
-				inputValue: this.state.inputValue
+				isPseudoFocused: this.state.isFocused && !this.props.multi
 			});
 		}
 		this.hasScrolledToOption = false;
-	},
+	}
 
 	handleInputFocus (event) {
 		if (this.props.disabled) return;
@@ -419,7 +322,7 @@ const Select = createClass({
 			isOpen: isOpen
 		});
 		this._openAfterFocus = false;
-	},
+	}
 
 	handleInputBlur (event) {
 		// The check for menu.contains(activeElement) is necessary to prevent IE11's scrollbar from closing the menu in certain contexts.
@@ -437,20 +340,16 @@ const Select = createClass({
 			isPseudoFocused: false,
 		};
 		if (this.props.onBlurResetsInput) {
-			onBlurredState.inputValue = '';
+			onBlurredState.inputValue = this.handleInputValueChange('');
 		}
 		this.setState(onBlurredState);
-	},
+	}
 
 	handleInputChange (event) {
 		let newInputValue = event.target.value;
 
-		if (this.state.inputValue !== event.target.value && this.props.onInputChange) {
-			let nextState = this.props.onInputChange(newInputValue);
-			// Note: != used deliberately here to catch undefined and null
-			if (nextState != null && typeof nextState !== 'object') {
-				newInputValue = '' + nextState;
-			}
+		if (this.state.inputValue !== event.target.value) {
+			newInputValue = this.handleInputValueChange(newInputValue);
 		}
 
 		this.setState({
@@ -458,7 +357,18 @@ const Select = createClass({
 			isPseudoFocused: false,
 			inputValue: newInputValue,
 		});
-	},
+	}
+
+	handleInputValueChange(newValue) {
+		if (this.props.onInputChange) {
+			let nextState = this.props.onInputChange(newValue);
+			// Note: != used deliberately here to catch undefined and null
+			if (nextState != null && typeof nextState !== 'object') {
+				newValue = '' + nextState;
+			}
+		}
+		return newValue;
+	}
 
 	handleKeyDown (event) {
 		if (this.props.disabled) return;
@@ -530,12 +440,12 @@ const Select = createClass({
 			default: return;
 		}
 		event.preventDefault();
-	},
+	}
 
 	handleValueClick (option, event) {
 		if (!this.props.onValueClick) return;
 		this.props.onValueClick(option, event);
-	},
+	}
 
 	handleMenuScroll (event) {
 		if (!this.props.onMenuScrollToBottom) return;
@@ -543,16 +453,16 @@ const Select = createClass({
 		if (target.scrollHeight > target.offsetHeight && !(target.scrollHeight - target.offsetHeight - target.scrollTop)) {
 			this.props.onMenuScrollToBottom();
 		}
-	},
+	}
 
 	handleRequired (value, multi) {
 		if (!value) return true;
 		return (multi ? value.length === 0 : Object.keys(value).length === 0);
-	},
+	}
 
 	getOptionLabel (op) {
 		return op[this.props.labelKey];
-	},
+	}
 
 	/**
 	 * Turns a value into an array from the given options
@@ -573,7 +483,7 @@ const Select = createClass({
 		}
 		var expandedValue = this.expandValue(value, props);
 		return expandedValue ? [expandedValue] : [];
-	},
+	}
 
 	/**
 	 * Retrieve a value from the given options and valueKey
@@ -588,7 +498,7 @@ const Select = createClass({
 		for (var i = 0; i < options.length; i++) {
 			if (options[i][valueKey] === value) return options[i];
 		}
-	},
+	}
 
 	setValue (value) {
 		if (this.props.autoBlur){
@@ -603,14 +513,15 @@ const Select = createClass({
 			value = this.props.multi ? value.map(i => i[this.props.valueKey]).join(this.props.delimiter) : value[this.props.valueKey];
 		}
 		this.props.onChange(value);
-	},
+	}
 
 	selectValue (value) {
 		//NOTE: update value in the callback to make sure the input value is empty so that there are no styling issues (Chrome had issue otherwise)
 		this.hasScrolledToOption = false;
 		if (this.props.multi) {
+			const updatedValue = this.props.onSelectResetsInput ? '' : this.state.inputValue;
 			this.setState({
-				inputValue: '',
+				inputValue: this.handleInputValueChange(updatedValue),
 				focusedIndex: null
 			}, () => {
 				this.addValue(value);
@@ -618,13 +529,13 @@ const Select = createClass({
 		} else {
 			this.setState({
 				isOpen: false,
-				inputValue: '',
+				inputValue: this.handleInputValueChange(''),
 				isPseudoFocused: this.state.isFocused,
 			}, () => {
 				this.setValue(value);
 			});
 		}
-	},
+	}
 
 	addValue (value) {
 		var valueArray = this.getValueArray(this.props.value);
@@ -638,20 +549,20 @@ const Select = createClass({
 			// focus the option below the selected one
 			this.focusOption(visibleOptions[lastValueIndex + 1]);
 		}
-	},
+	}
 
 	popValue () {
 		var valueArray = this.getValueArray(this.props.value);
 		if (!valueArray.length) return;
 		if (valueArray[valueArray.length-1].clearableValue === false) return;
-		this.setValue(valueArray.slice(0, valueArray.length - 1));
-	},
+		this.setValue(this.props.multi ? valueArray.slice(0, valueArray.length - 1) : null);
+	}
 
 	removeValue (value) {
 		var valueArray = this.getValueArray(this.props.value);
 		this.setValue(valueArray.filter(i => i !== value));
 		this.focus();
-	},
+	}
 
 	clearValue (event) {
 		// if the event was triggered by a mousedown and not the primary
@@ -664,9 +575,9 @@ const Select = createClass({
 		this.setValue(this.getResetValue());
 		this.setState({
 			isOpen: false,
-			inputValue: '',
+			inputValue: this.handleInputValueChange(''),
 		}, this.focus);
-	},
+	}
 
 	getResetValue () {
 		if (this.props.resetValue !== undefined) {
@@ -676,37 +587,37 @@ const Select = createClass({
 		} else {
 			return null;
 		}
-	},
+	}
 
 	focusOption (option) {
 		this.setState({
 			focusedOption: option
 		});
-	},
+	}
 
 	focusNextOption () {
 		this.focusAdjacentOption('next');
-	},
+	}
 
 	focusPreviousOption () {
 		this.focusAdjacentOption('previous');
-	},
+	}
 
 	focusPageUpOption () {
 		this.focusAdjacentOption('page_up');
-	},
+	}
 
 	focusPageDownOption () {
 		this.focusAdjacentOption('page_down');
-	},
+	}
 
 	focusStartOption () {
 		this.focusAdjacentOption('start');
-	},
+	}
 
 	focusEndOption () {
 		this.focusAdjacentOption('end');
-	},
+	}
 
 	focusAdjacentOption (dir) {
 		var options = this._visibleOptions
@@ -765,21 +676,21 @@ const Select = createClass({
 			focusedIndex: options[focusedIndex].index,
 			focusedOption: options[focusedIndex].option
 		});
-	},
+	}
 
 	getFocusedOption () {
 		return this._focusedOption;
-	},
+	}
 
 	getInputValue () {
 		return this.state.inputValue;
-	},
+	}
 
 	selectFocusedOption () {
 		if (this._focusedOption) {
 			return this.selectValue(this._focusedOption);
 		}
-	},
+	}
 
 	renderLoading () {
 		if (!this.props.isLoading) return;
@@ -788,7 +699,7 @@ const Select = createClass({
 				<span className="Select-loading" />
 			</span>
 		);
-	},
+	}
 
 	renderValue (valueArray, isOpen) {
 		let renderLabel = this.props.valueRenderer || this.getOptionLabel;
@@ -828,7 +739,7 @@ const Select = createClass({
 				</ValueComponent>
 			);
 		}
-	},
+	}
 
 	renderInput (valueArray, focusedOptionIndex) {
 		var className = classNames('Select-input', this.props.inputProps.className);
@@ -900,7 +811,7 @@ const Select = createClass({
 				<input {...inputProps} />
 			</div>
 		);
-	},
+	}
 
 	renderClear () {
 
@@ -918,11 +829,11 @@ const Select = createClass({
 				{clear}
 			</span>
 		);
-	},
+	}
 
 	renderArrow () {
 		const onMouseDown = this.handleMouseDownOnArrow;
-                const isOpen = this.state.isOpen;
+		const isOpen = this.state.isOpen;
 		const arrow = this.props.arrowRenderer({ onMouseDown, isOpen });
 
 		return (
@@ -933,7 +844,7 @@ const Select = createClass({
 				{arrow}
 			</span>
 		);
-	},
+	}
 
 	filterOptions (excludeOptions) {
 		var filterValue = this.state.inputValue;
@@ -961,13 +872,13 @@ const Select = createClass({
 		} else {
 			return options;
 		}
-	},
+	}
 
 	onOptionRef(ref, isFocused) {
 		if (isFocused) {
 			this.focused = ref;
 		}
-	},
+	}
 
 	renderMenu (options, valueArray, focusedOption) {
 		if (options && options.length) {
@@ -996,7 +907,7 @@ const Select = createClass({
 		} else {
 			return null;
 		}
-	},
+	}
 
 	renderHiddenField (valueArray) {
 		if (!this.props.name) return;
@@ -1019,7 +930,7 @@ const Select = createClass({
 				value={stringifyValue(item[this.props.valueKey])}
 				disabled={this.props.disabled} />
 		));
-	},
+	}
 
 	getFocusableOptionIndex (selectedOption) {
 		var options = this._visibleOptions;
@@ -1045,7 +956,7 @@ const Select = createClass({
 			if (!options[i].disabled) return i;
 		}
 		return null;
-	},
+	}
 
 	renderOuter (options, valueArray, focusedOption) {
 		let menu = this.renderMenu(options, valueArray, focusedOption);
@@ -1055,7 +966,7 @@ const Select = createClass({
 
 		return (
 			<div ref={ref => this.menuContainer = ref} className="Select-menu-outer" style={this.props.menuContainerStyle}>
-				<div ref={ref => this.menu = ref} role="listbox" className="Select-menu" id={this._instancePrefix + '-list'}
+				<div ref={ref => this.menu = ref} role="listbox" tabIndex={-1} className="Select-menu" id={this._instancePrefix + '-list'}
 						 style={this.props.menuStyle}
 						 onScroll={this.handleMenuScroll}
 						 onMouseDown={this.handleMouseDownOnMenu}>
@@ -1063,7 +974,7 @@ const Select = createClass({
 				</div>
 			</div>
 		);
-	},
+	}
 
 	render () {
 		let valueArray = this.getValueArray(this.props.value);
@@ -1132,7 +1043,122 @@ const Select = createClass({
 			</div>
 		);
 	}
+};
 
-});
+Select.propTypes = {
+    'aria-describedby': PropTypes.string, // HTML ID(s) of element(s) that should be used to describe this input (for assistive tech)
+    'aria-label': PropTypes.string,       // Aria label (for assistive tech)
+    'aria-labelledby': PropTypes.string,  // HTML ID of an element that should be used as the label (for assistive tech)
+		addLabelText: PropTypes.string,       // placeholder displayed when you want to add a label on a multi-value input
+    arrowRenderer: PropTypes.func,        // Create drop-down caret element
+    autoBlur: PropTypes.bool,             // automatically blur the component when an option is selected
+    autofocus: PropTypes.bool,            // autofocus the component on mount
+    autosize: PropTypes.bool,             // whether to enable autosizing or not
+    backspaceRemoves: PropTypes.bool,     // whether backspace removes an item if there is no text input
+    backspaceToRemoveMessage: PropTypes.string,  // Message to use for screenreaders to press backspace to remove the current item - {label} is replaced with the item label
+    className: PropTypes.string,          // className for the outer element
+    clearAllText: stringOrNode,           // title for the "clear" control when multi: true
+    clearRenderer: PropTypes.func,        // create clearable x element
+    clearValueText: stringOrNode,         // title for the "clear" control
+    clearable: PropTypes.bool,            // should it be possible to reset value
+    deleteRemoves: PropTypes.bool,        // whether backspace removes an item if there is no text input
+    delimiter: PropTypes.string,          // delimiter to use to join multiple values for the hidden field value
+    disabled: PropTypes.bool,             // whether the Select is disabled or not
+    escapeClearsValue: PropTypes.bool,    // whether escape clears the value when the menu is closed
+    filterOption: PropTypes.func,         // method to filter a single option (option, filterString)
+    filterOptions: PropTypes.any,         // boolean to enable default filtering or function to filter the options array ([options], filterString, [values])
+    ignoreAccents: PropTypes.bool,        // whether to strip diacritics when filtering
+    ignoreCase: PropTypes.bool,           // whether to perform case-insensitive filtering
+    inputProps: PropTypes.object,         // custom attributes for the Input
+    inputRenderer: PropTypes.func,        // returns a custom input component
+    instanceId: PropTypes.string,         // set the components instanceId
+    isLoading: PropTypes.bool,            // whether the Select is loading externally or not (such as options being loaded)
+    joinValues: PropTypes.bool,           // joins multiple values into a single form field with the delimiter (legacy mode)
+    labelKey: PropTypes.string,           // path of the label value in option objects
+    matchPos: PropTypes.string,           // (any|start) match the start or entire string when filtering
+    matchProp: PropTypes.string,          // (any|label|value) which option property to filter on
+    menuBuffer: PropTypes.number,         // optional buffer (in px) between the bottom of the viewport and the bottom of the menu
+    menuContainerStyle: PropTypes.object, // optional style to apply to the menu container
+    menuRenderer: PropTypes.func,         // renders a custom menu with options
+    menuStyle: PropTypes.object,          // optional style to apply to the menu
+    multi: PropTypes.bool,                // multi-value input
+    name: PropTypes.string,               // generates a hidden <input /> tag with this field name for html forms
+    noResultsText: stringOrNode,          // placeholder displayed when there are no matching search results
+    onBlur: PropTypes.func,               // onBlur handler: function (event) {}
+    onBlurResetsInput: PropTypes.bool,    // whether input is cleared on blur
+    onChange: PropTypes.func,             // onChange handler: function (newValue) {}
+    onClose: PropTypes.func,              // fires when the menu is closed
+    onCloseResetsInput: PropTypes.bool,   // whether input is cleared when menu is closed through the arrow
+    onFocus: PropTypes.func,              // onFocus handler: function (event) {}
+    onInputChange: PropTypes.func,        // onInputChange handler: function (inputValue) {}
+    onInputKeyDown: PropTypes.func,       // input keyDown handler: function (event) {}
+    onMenuScrollToBottom: PropTypes.func, // fires when the menu is scrolled to the bottom; can be used to paginate options
+    onOpen: PropTypes.func,               // fires when the menu is opened
+    onSelectResetsInput: PropTypes.bool,  // whether input is cleared on select (works only for multiselect)
+    onValueClick: PropTypes.func,         // onClick handler for value labels: function (value, event) {}
+    openAfterFocus: PropTypes.bool,       // boolean to enable opening dropdown when focused
+    openOnFocus: PropTypes.bool,          // always open options menu on focus
+    optionClassName: PropTypes.string,    // additional class(es) to apply to the <Option /> elements
+    optionComponent: PropTypes.func,      // option component to render in dropdown
+    optionRenderer: PropTypes.func,       // optionRenderer: function (option) {}
+    options: PropTypes.array,             // array of options
+    pageSize: PropTypes.number,           // number of entries to page when using page up/down keys
+    placeholder: stringOrNode,            // field placeholder, displayed when there's no value
+    required: PropTypes.bool,             // applies HTML5 required attribute when needed
+    resetValue: PropTypes.any,            // value to use when you clear the control
+    scrollMenuIntoView: PropTypes.bool,   // boolean to enable the viewport to shift so that the full menu fully visible when engaged
+    searchable: PropTypes.bool,           // whether to enable searching feature or not
+    simpleValue: PropTypes.bool,          // pass the value to onChange as a simple value (legacy pre 1.0 mode), defaults to false
+    style: PropTypes.object,              // optional style to apply to the control
+    tabIndex: PropTypes.string,           // optional tab index of the control
+    tabSelectsValue: PropTypes.bool,      // whether to treat tabbing out while focused to be value selection
+    value: PropTypes.any,                 // initial field value
+    valueComponent: PropTypes.func,       // value component to render
+    valueKey: PropTypes.string,           // path of the label value in option objects
+    valueRenderer: PropTypes.func,        // valueRenderer: function (option) {}
+    wrapperStyle: PropTypes.object,       // optional style to apply to the component wrapper
+};
+
+Select.defaultProps = {
+    addLabelText: 'Add "{label}"?',
+    arrowRenderer: defaultArrowRenderer,
+    autosize: true,
+    backspaceRemoves: true,
+    backspaceToRemoveMessage: 'Press backspace to remove {label}',
+    clearable: true,
+    clearAllText: 'Clear all',
+    clearRenderer: defaultClearRenderer,
+    clearValueText: 'Clear value',
+    deleteRemoves: true,
+    delimiter: ',',
+    disabled: false,
+    escapeClearsValue: true,
+    filterOptions: defaultFilterOptions,
+    ignoreAccents: true,
+    ignoreCase: true,
+    inputProps: {},
+    isLoading: false,
+    joinValues: false,
+    labelKey: 'label',
+    matchPos: 'any',
+    matchProp: 'any',
+    menuBuffer: 0,
+    menuRenderer: defaultMenuRenderer,
+    multi: false,
+    noResultsText: 'No results found',
+    onBlurResetsInput: true,
+    onSelectResetsInput: true,
+    onCloseResetsInput: true,
+    optionComponent: Option,
+    pageSize: 5,
+    placeholder: 'Select...',
+    required: false,
+    scrollMenuIntoView: true,
+    searchable: true,
+    simpleValue: false,
+    tabSelectsValue: true,
+    valueComponent: Value,
+    valueKey: 'value',
+};
 
 export default Select;
