@@ -35,4 +35,4 @@ function filterOptions (options, filterValue, excludeOptions, props) {
 	});
 }
 
-module.exports = filterOptions;
+export default filterOptions;
