@@ -1,64 +1,34 @@
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-	value: true
-});
-
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
-var _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ('value' in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import Select from './Select';
+import stripDiacritics from './utils/stripDiacritics';
+const propTypes = {
+	autoload: PropTypes.bool.isRequired, // automatically call the `loadOptions` prop on-mount; defaults to true
+	cache: PropTypes.any, // object to use to cache results; set to null/false to disable caching
+	children: PropTypes.func.isRequired, // Child function responsible for creating the inner Select component; (props: Object): PropTypes.element
+	ignoreAccents: PropTypes.bool, // strip diacritics when filtering; defaults to true
+	ignoreCase: PropTypes.bool, // perform case-insensitive filtering; defaults to true
+	loadOptions: PropTypes.func.isRequired, // callback to load options asynchronously; (inputValue: string, callback: Function): ?Promise
+	loadingPlaceholder: PropTypes.oneOfType([// replaces the placeholder while options are loading
+	PropTypes.string, PropTypes.node]),
+	multi: PropTypes.bool, // multi-value input
+	noResultsText: PropTypes.oneOfType([// field noResultsText, displayed when no options come back from the server
+	PropTypes.string, PropTypes.node]),
+	onChange: PropTypes.func, // onChange handler: function (newValue) {}
+	onInputChange: PropTypes.func, // optional for keeping track of what is being typed
+	options: PropTypes.array.isRequired, // array of options
+	placeholder: PropTypes.oneOfType([// field placeholder, displayed when there's no value (shared with Select)
+	PropTypes.string, PropTypes.node]),
+	searchPromptText: PropTypes.oneOfType([// label to prompt for search input
+	PropTypes.string, PropTypes.node]),
+	value: PropTypes.any // initial field value
+};
 
-var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_again) { var object = _x, property = _x2, receiver = _x3; _again = false; if (object === null) object = Function.prototype; var desc = Object.getOwnPropertyDescriptor(object, property); if (desc === undefined) { var parent = Object.getPrototypeOf(object); if (parent === null) { return undefined; } else { _x = parent; _x2 = property; _x3 = receiver; _again = true; desc = parent = undefined; continue _function; } } else if ('value' in desc) { return desc.value; } else { var getter = desc.get; if (getter === undefined) { return undefined; } return getter.call(receiver); } } };
+const defaultCache = {};
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
-
-function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-var _propTypes = require('prop-types');
-
-var _propTypes2 = _interopRequireDefault(_propTypes);
-
-var _Select = require('./Select');
-
-var _Select2 = _interopRequireDefault(_Select);
-
-var _utilsStripDiacritics = require('./utils/stripDiacritics');
-
-var _utilsStripDiacritics2 = _interopRequireDefault(_utilsStripDiacritics);
-
-var propTypes = {
-	autoload: _propTypes2['default'].bool.isRequired, // automatically call the `loadOptions` prop on-mount; defaults to true
-	cache: _propTypes2['default'].any, // object to use to cache results; set to null/false to disable caching
-	children: _propTypes2['default'].func.isRequired, // Child function responsible for creating the inner Select component; (props: Object): PropTypes.element
-	ignoreAccents: _propTypes2['default'].bool, // strip diacritics when filtering; defaults to true
-	ignoreCase: _propTypes2['default'].bool, // perform case-insensitive filtering; defaults to true
-	loadingPlaceholder: _propTypes2['default'].oneOfType([// replaces the placeholder while options are loading
-	_propTypes2['default'].string, _propTypes2['default'].node]),
-	loadOptions: _propTypes2['default'].func.isRequired, // callback to load options asynchronously; (inputValue: string, callback: Function): ?Promise
-	multi: _propTypes2['default'].bool, // multi-value input
-	options: _propTypes2['default'].array.isRequired, // array of options
-	placeholder: _propTypes2['default'].oneOfType([// field placeholder, displayed when there's no value (shared with Select)
-	_propTypes2['default'].string, _propTypes2['default'].node]),
-	noResultsText: _propTypes2['default'].oneOfType([// field noResultsText, displayed when no options come back from the server
-	_propTypes2['default'].string, _propTypes2['default'].node]),
-	onChange: _propTypes2['default'].func, // onChange handler: function (newValue) {}
-	searchPromptText: _propTypes2['default'].oneOfType([// label to prompt for search input
-	_propTypes2['default'].string, _propTypes2['default'].node]),
-	onInputChange: _propTypes2['default'].func, // optional for keeping track of what is being typed
-	value: _propTypes2['default'].any };
-
-// initial field value
-var defaultCache = {};
-
-var defaultProps = {
+const defaultProps = {
 	autoload: true,
 	cache: defaultCache,
 	children: defaultChildren,
@@ -69,13 +39,9 @@ var defaultProps = {
 	searchPromptText: 'Type to search'
 };
 
-var Async = (function (_Component) {
-	_inherits(Async, _Component);
-
-	function Async(props, context) {
-		_classCallCheck(this, Async);
-
-		_get(Object.getPrototypeOf(Async.prototype), 'constructor', this).call(this, props, context);
+export default class Async extends Component {
+	constructor(props, context) {
+		super(props, context);
 
 		this._cache = props.cache === defaultCache ? {} : props.cache;
 
@@ -87,184 +53,145 @@ var Async = (function (_Component) {
 		this._onInputChange = this._onInputChange.bind(this);
 	}
 
-	_createClass(Async, [{
-		key: 'componentDidMount',
-		value: function componentDidMount() {
-			var autoload = this.props.autoload;
+	componentDidMount() {
+		const { autoload } = this.props;
 
-			if (autoload) {
-				this.loadOptions('');
-			}
+		if (autoload) {
+			this.loadOptions('');
 		}
-	}, {
-		key: 'componentWillUpdate',
-		value: function componentWillUpdate(nextProps, nextState) {
-			var _this = this;
+	}
 
-			var propertiesToSync = ['options'];
-			propertiesToSync.forEach(function (prop) {
-				if (_this.props[prop] !== nextProps[prop]) {
-					_this.setState(_defineProperty({}, prop, nextProps[prop]));
-				}
+	componentWillReceiveProps(nextProps) {
+		if (nextProps.options !== this.props.options) {
+			this.setState({
+				options: nextProps.options
 			});
 		}
-	}, {
-		key: 'clearOptions',
-		value: function clearOptions() {
-			this.setState({ options: [] });
+	}
+
+	clearOptions() {
+		this.setState({ options: [] });
+	}
+
+	loadOptions(inputValue) {
+		const { loadOptions } = this.props;
+		const cache = this._cache;
+
+		if (cache && cache.hasOwnProperty(inputValue)) {
+			this.setState({
+				options: cache[inputValue]
+			});
+
+			return;
 		}
-	}, {
-		key: 'loadOptions',
-		value: function loadOptions(inputValue) {
-			var _this2 = this;
 
-			var loadOptions = this.props.loadOptions;
+		const callback = (error, data) => {
+			if (callback === this._callback) {
+				this._callback = null;
 
-			var cache = this._cache;
+				const options = data && data.options || [];
 
-			if (cache && cache.hasOwnProperty(inputValue)) {
-				this.setState({
-					options: cache[inputValue]
-				});
-
-				return;
-			}
-
-			var callback = function callback(error, data) {
-				if (callback === _this2._callback) {
-					_this2._callback = null;
-
-					var options = data && data.options || [];
-
-					if (cache) {
-						cache[inputValue] = options;
-					}
-
-					_this2.setState({
-						isLoading: false,
-						options: options
-					});
+				if (cache) {
+					cache[inputValue] = options;
 				}
-			};
 
-			// Ignore all but the most recent request
-			this._callback = callback;
-
-			var promise = loadOptions(inputValue, callback);
-			if (promise) {
-				promise.then(function (data) {
-					return callback(null, data);
-				}, function (error) {
-					return callback(error);
-				});
-			}
-
-			if (this._callback && !this.state.isLoading) {
 				this.setState({
-					isLoading: true
+					isLoading: false,
+					options
 				});
 			}
+		};
 
-			return inputValue;
+		// Ignore all but the most recent request
+		this._callback = callback;
+
+		const promise = loadOptions(inputValue, callback);
+		if (promise) {
+			promise.then(data => callback(null, data), error => callback(error));
 		}
-	}, {
-		key: '_onInputChange',
-		value: function _onInputChange(inputValue) {
-			var _props = this.props;
-			var ignoreAccents = _props.ignoreAccents;
-			var ignoreCase = _props.ignoreCase;
-			var onInputChange = _props.onInputChange;
 
-			if (ignoreAccents) {
-				inputValue = (0, _utilsStripDiacritics2['default'])(inputValue);
-			}
-
-			if (ignoreCase) {
-				inputValue = inputValue.toLowerCase();
-			}
-
-			if (onInputChange) {
-				onInputChange(inputValue);
-			}
-
-			return this.loadOptions(inputValue);
+		if (this._callback && !this.state.isLoading) {
+			this.setState({
+				isLoading: true
+			});
 		}
-	}, {
-		key: 'inputValue',
-		value: function inputValue() {
-			if (this.select) {
-				return this.select.state.inputValue;
-			}
-			return '';
+	}
+
+	_onInputChange(inputValue) {
+		const { ignoreAccents, ignoreCase, onInputChange } = this.props;
+		let transformedInputValue = inputValue;
+
+		if (ignoreAccents) {
+			transformedInputValue = stripDiacritics(transformedInputValue);
 		}
-	}, {
-		key: 'noResultsText',
-		value: function noResultsText() {
-			var _props2 = this.props;
-			var loadingPlaceholder = _props2.loadingPlaceholder;
-			var noResultsText = _props2.noResultsText;
-			var searchPromptText = _props2.searchPromptText;
-			var isLoading = this.state.isLoading;
 
-			var inputValue = this.inputValue();
-
-			if (isLoading) {
-				return loadingPlaceholder;
-			}
-			if (inputValue && noResultsText) {
-				return noResultsText;
-			}
-			return searchPromptText;
+		if (ignoreCase) {
+			transformedInputValue = transformedInputValue.toLowerCase();
 		}
-	}, {
-		key: 'focus',
-		value: function focus() {
-			this.select.focus();
+
+		if (onInputChange) {
+			onInputChange(transformedInputValue);
 		}
-	}, {
-		key: 'render',
-		value: function render() {
-			var _this3 = this;
 
-			var _props3 = this.props;
-			var children = _props3.children;
-			var loadingPlaceholder = _props3.loadingPlaceholder;
-			var placeholder = _props3.placeholder;
-			var _state = this.state;
-			var isLoading = _state.isLoading;
-			var options = _state.options;
+		this.loadOptions(transformedInputValue);
 
-			var props = {
-				noResultsText: this.noResultsText(),
-				placeholder: isLoading ? loadingPlaceholder : placeholder,
-				options: isLoading && loadingPlaceholder ? [] : options,
-				ref: function ref(_ref) {
-					return _this3.select = _ref;
-				},
-				onChange: function onChange(newValues) {
-					if (_this3.props.multi && _this3.props.value && newValues.length > _this3.props.value.length) {
-						_this3.clearOptions();
-					}
-					_this3.props.onChange(newValues);
+		// Return the original input value to avoid modifying the user's view of the input while typing.
+		return inputValue;
+	}
+
+	inputValue() {
+		if (this.select) {
+			return this.select.state.inputValue;
+		}
+		return '';
+	}
+
+	noResultsText() {
+		const { loadingPlaceholder, noResultsText, searchPromptText } = this.props;
+		const { isLoading } = this.state;
+
+		const inputValue = this.inputValue();
+
+		if (isLoading) {
+			return loadingPlaceholder;
+		}
+		if (inputValue && noResultsText) {
+			return noResultsText;
+		}
+		return searchPromptText;
+	}
+
+	focus() {
+		this.select.focus();
+	}
+
+	render() {
+		const { children, loadingPlaceholder, placeholder } = this.props;
+		const { isLoading, options } = this.state;
+
+		const props = {
+			noResultsText: this.noResultsText(),
+			placeholder: isLoading ? loadingPlaceholder : placeholder,
+			options: isLoading && loadingPlaceholder ? [] : options,
+			ref: ref => this.select = ref,
+			onChange: newValues => {
+				if (this.props.multi && this.props.value && newValues.length > this.props.value.length) {
+					this.clearOptions();
 				}
-			};
+				this.props.onChange(newValues);
+			}
+		};
 
-			return children(_extends({}, this.props, props, {
-				isLoading: isLoading,
-				onInputChange: this._onInputChange
-			}));
-		}
-	}]);
-
-	return Async;
-})(_react.Component);
-
-exports['default'] = Async;
+		return children(_extends({}, this.props, props, {
+			isLoading,
+			onInputChange: this._onInputChange
+		}));
+	}
+}
 
 Async.propTypes = propTypes;
 Async.defaultProps = defaultProps;
 
 function defaultChildren(props) {
-	return _react2['default'].createElement(_Select2['default'], props);
+	return React.createElement(Select, props);
 }
-module.exports = exports['default'];
