@@ -1,38 +1,22 @@
-'use strict';
+import React from 'react';
+import PropTypes from 'prop-types';
+import classNames from 'classnames';
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+class Option extends React.Component {
 
-var _react = require('react');
+	constructor(props) {
+		super(props);
 
-var _react2 = _interopRequireDefault(_react);
+		this.handleMouseDown = this.handleMouseDown.bind(this);
+		this.handleMouseEnter = this.handleMouseEnter.bind(this);
+		this.handleMouseMove = this.handleMouseMove.bind(this);
+		this.handleTouchStart = this.handleTouchStart.bind(this);
+		this.handleTouchEnd = this.handleTouchEnd.bind(this);
+		this.handleTouchMove = this.handleTouchMove.bind(this);
+		this.onFocus = this.onFocus.bind(this);
+	}
 
-var _createReactClass = require('create-react-class');
-
-var _createReactClass2 = _interopRequireDefault(_createReactClass);
-
-var _propTypes = require('prop-types');
-
-var _propTypes2 = _interopRequireDefault(_propTypes);
-
-var _classnames = require('classnames');
-
-var _classnames2 = _interopRequireDefault(_classnames);
-
-var Option = (0, _createReactClass2['default'])({
-	propTypes: {
-		children: _propTypes2['default'].node,
-		className: _propTypes2['default'].string, // className (based on mouse position)
-		instancePrefix: _propTypes2['default'].string.isRequired, // unique prefix for the ids (used for aria)
-		isDisabled: _propTypes2['default'].bool, // the option is disabled
-		isFocused: _propTypes2['default'].bool, // the option is focused
-		isSelected: _propTypes2['default'].bool, // the option is selected
-		onFocus: _propTypes2['default'].func, // method to handle mouseEnter on option element
-		onSelect: _propTypes2['default'].func, // method to handle click on option element
-		onUnfocus: _propTypes2['default'].func, // method to handle mouseLeave on option element
-		option: _propTypes2['default'].object.isRequired, // object that is base for that option
-		optionIndex: _propTypes2['default'].number },
-	// index of the option, used to generate unique ids for aria
-	blockEvent: function blockEvent(event) {
+	blockEvent(event) {
 		event.preventDefault();
 		event.stopPropagation();
 		if (event.target.tagName !== 'A' || !('href' in event.target)) {
@@ -43,60 +27,57 @@ var Option = (0, _createReactClass2['default'])({
 		} else {
 			window.location.href = event.target.href;
 		}
-	},
+	}
 
-	handleMouseDown: function handleMouseDown(event) {
+	handleMouseDown(event) {
 		event.preventDefault();
 		event.stopPropagation();
 		this.props.onSelect(this.props.option, event);
-	},
+	}
 
-	handleMouseEnter: function handleMouseEnter(event) {
+	handleMouseEnter(event) {
 		this.onFocus(event);
-	},
+	}
 
-	handleMouseMove: function handleMouseMove(event) {
+	handleMouseMove(event) {
 		this.onFocus(event);
-	},
+	}
 
-	handleTouchEnd: function handleTouchEnd(event) {
+	handleTouchEnd(event) {
 		// Check if the view is being dragged, In this case
 		// we don't want to fire the click event (because the user only wants to scroll)
 		if (this.dragging) return;
 
 		this.handleMouseDown(event);
-	},
+	}
 
-	handleTouchMove: function handleTouchMove(event) {
+	handleTouchMove(event) {
 		// Set a flag that the view is being dragged
 		this.dragging = true;
-	},
+	}
 
-	handleTouchStart: function handleTouchStart(event) {
+	handleTouchStart(event) {
 		// Set a flag that the view is not being dragged
 		this.dragging = false;
-	},
+	}
 
-	onFocus: function onFocus(event) {
+	onFocus(event) {
 		if (!this.props.isFocused) {
 			this.props.onFocus(this.props.option, event);
 		}
-	},
-	render: function render() {
-		var _props = this.props;
-		var option = _props.option;
-		var instancePrefix = _props.instancePrefix;
-		var optionIndex = _props.optionIndex;
+	}
 
-		var className = (0, _classnames2['default'])(this.props.className, option.className);
+	render() {
+		var { option, instancePrefix, optionIndex } = this.props;
+		var className = classNames(this.props.className, option.className);
 
-		return option.disabled ? _react2['default'].createElement(
+		return option.disabled ? React.createElement(
 			'div',
 			{ className: className,
 				onMouseDown: this.blockEvent,
 				onClick: this.blockEvent },
 			this.props.children
-		) : _react2['default'].createElement(
+		) : React.createElement(
 			'div',
 			{ className: className,
 				style: option.style,
@@ -112,6 +93,20 @@ var Option = (0, _createReactClass2['default'])({
 			this.props.children
 		);
 	}
-});
+};
 
-module.exports = Option;
+Option.propTypes = {
+	children: PropTypes.node,
+	className: PropTypes.string, // className (based on mouse position)
+	instancePrefix: PropTypes.string.isRequired, // unique prefix for the ids (used for aria)
+	isDisabled: PropTypes.bool, // the option is disabled
+	isFocused: PropTypes.bool, // the option is focused
+	isSelected: PropTypes.bool, // the option is selected
+	onFocus: PropTypes.func, // method to handle mouseEnter on option element
+	onSelect: PropTypes.func, // method to handle click on option element
+	onUnfocus: PropTypes.func, // method to handle mouseLeave on option element
+	option: PropTypes.object.isRequired, // object that is base for that option
+	optionIndex: PropTypes.number // index of the option, used to generate unique ids for aria
+};
+
+export default Option;
