@@ -1,24 +1,13 @@
-"use strict";
+import React from 'react';
+import PropTypes from 'prop-types';
 
-Object.defineProperty(exports, "__esModule", {
-	value: true
-});
-exports["default"] = arrowRenderer;
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-function arrowRenderer(_ref) {
-	var onMouseDown = _ref.onMouseDown;
-
-	return _react2["default"].createElement("span", {
-		className: "Select-arrow",
+export default function arrowRenderer({ onMouseDown }) {
+	return React.createElement('span', {
+		className: 'Select-arrow',
 		onMouseDown: onMouseDown
 	});
-}
+};
 
-;
-module.exports = exports["default"];
+arrowRenderer.propTypes = {
+	onMouseDown: PropTypes.func
+};
