@@ -1,22 +1,8 @@
-'use strict';
+import React from 'react';
 
-Object.defineProperty(exports, '__esModule', {
-	value: true
-});
-exports['default'] = clearRenderer;
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-var _react = require('react');
-
-var _react2 = _interopRequireDefault(_react);
-
-function clearRenderer() {
-	return _react2['default'].createElement('span', {
+export default function clearRenderer() {
+	return React.createElement('span', {
 		className: 'Select-clear',
 		dangerouslySetInnerHTML: { __html: '&times;' }
 	});
-}
-
-;
-module.exports = exports['default'];
+};
