@@ -16,11 +16,11 @@ var expect = unexpected
 
 var React = require('react');
 var ReactDOM = require('react-dom');
-var TestUtils = require('react-addons-test-utils');
+var TestUtils = require('react-dom/test-utils');
 
 var OPTION = { label: 'TEST-LABEL', value: 'TEST-VALUE' };
 
-var Value = require('../src/Value');
+var Value = require('../src/Value').default;
 
 describe('Value component', function() {
 
